systemd: check multi-user.target instead of default.target
The behavior of `default.target` is a bit special and doesn't report as
active consistently even though it is an alias for `multi-user.target`
which is indeed active. Not really worth any more of our time figuring
out why this is the case so just switch the check.

Required for systemd 217.