kola: move rolling upgrade test to subcommand
This was pulled out from other automated kola tests because it is not
suited for automatic testing until supplying the paths of pre-compiled
binaries is no longer necessary. It is also not well-suited for
automated kola testing as is because it tests software versions that are
independant of the OS release process.