kola/tests/kubernetes: fallback to /usr/bin/kubelet not built-in wrapper
This test would attempt to fallback to placing a kubelet-wrapper on disk
if the CoreOS image didn't have one built in. However, the wrapper
assumes rkt fly is available. In practice almost no CoreOS releases
except some stables versions have rkt fly but not the wrapper. Instead
fallback to /usr/bin/kubelet so this test can pass on more images
including the current stable and beta as of the time of this commit.