repo: verify trees are in sync with release manifest
This will serve as a sanity check for builds, making sure that all git
trees are actually checked out to the revision specified in a release
manifest file. This is to counter the fact that I know it is possible to
get `repo sync` to stop updating working trees but I don't know how to
reproduce it. Instead of adding extra goo like `git reset --hard HEAD`
like many automated users of repo and hoping for the best this should
catch repo in the act so we actually know what its limits are first.