index: do not store redirect pages in general objects list
Keeping with normal HTTP server behavior, directory paths without the
trailing / should redirect, appending / to the path. There is nothing
special about these objects to reliably indicate that they are redirects
other than happening to conflict with the prefix of another object so
the parent Directory has to detect them.