index: use a hash to trigger index updates instead of time
This is more robust since it can detect deletion of child nodes and
changes to the html template. For now CRC32c is used, if that is
insufficient we can switch to MD5 in the future. As an added bonus,
the storage server validates the upload using the provided hash.