index: add index page generation test and benchmark
I spent some time tinkering with how fast index page generation was now
that it had to be invoked just to detect if an update is needed.
Potentially that could slow down re-indexing very large buckets even
when only a few pages need to be updated. The result:

    $ go test -bench .
    PASS
    BenchmarkBuildIndex-8          10000        163589 ns/op
    BenchmarkRawBuildIndex-8      300000          5081 ns/op
    ok      github.com/coreos/mantle/index  3.235s

So the answer is yes, even with the experimental code being more-correct
by properly escaping text which the old code doesn't do. However
fetching the object list for a large bucket is still so slow that the
few seconds vs no seconds of these options doesn't matter. So I'm
leaving it as-is for now but leaving the benchmark in case we want to
switch later.