build: support incremental builds
Without the -i option, go build and test do not save any of the built
packages for later re-use. This results in a hell of a lot of
duplication. For me this reduces a single run of the build script from
50 seconds to 24, then 14 seconds after that. The time remaining is how
long it takes to link all of the commands, go build won't skip that.