ore+plume: replace cloud/storage with api/storage package.
The cloud/storage package is documented as experimental and has been
significantly refactored since this code was written. Since this code
needs to be reworked anyway may as well switch to the somewhat more
stable api/storage package.

There is a far amount of code duplication in these commands but fixing
that is left for another day.