storage/index: use sorted slices, drop extra param structs
Commit 43340a2c03 added a bunch of escaping/processing of data passed to
the template. I forgot that I was using html/template, which does that
automatically. Also it is easy to trim path names in the template so
stop doing that in advance with maps. Now that we have slices the data
can just start off sorted how we want it.

And hey look, I'm terrible at HTML too. Fix the <br/> tags.