storage: add support disabling recursive fetching/syncing/indexing
When copying a directory to an indexed bucket it would be nice to update
the parent directory's index without having to fetch and re-index the
entire existing tree. There is one quirk in this mode of operation
though: doing a shallow fetch means we cannot check if subdirectories
are empty (contain only index objects but nothing else). In practice
this isn't likely to come up being a non-obvious difference in behavior
it's worth calling out. The non-recursive syncing option doesn't have an
immediate need but added it anyway to keep features in sync.