plume: add new index command based on the release specs
This makes it trivial to re-index any portion of the download sites,
simply by specifying the appropriate channel, board, and/or version. The
lower level functionality of indexing by URL is in ore now.

This logic is a little silly looking given that the spec definition
structure was set up for processing a single release rather than many at
once but it works well enough.