system: switch from special AnonFile type to plain os.File
After some reflection and coding the use of a special type is a bit
awkward. Code doesn't care about the extra Link method still has to deal
with the new type. Instead turn Link into the function LinkFile which
has the nice property of being similar to the standard os.Link.