system: add test to cover LinkFile's behavior with normal files
Files opened with O_TMPFILE (but without O_EXCL) are a special case with
regard to linking, they may be linked into a filesystem when there are
no other links to them. Normally files cannot be linked after link count
drops to 0 so add a test to cover/document this different behavior.