all: refactor repo layout a bit
The idea here is that the top level directory will contain domain types
such that reading these types will give people a good idea of how tests
function in this repo. Eventually we may want to mock more kola types to
make tests easier for this domain.

I figured I'd start be separating kubernetes cluster setup from the
tests that run against them via the Cluster type. Tests should still be
aware of the underlying setup method so its fine that say a bootkube
test is aware that the cluster is self-hosted and that down the road a
tectonic test would understand that the Cluster contains Tectonic
specific components.