kola/tests/fleet: migrate test configs to Ignition
Since fleetd has the annoying behavior of not acknowledging that
etcd is ready until a few seconds after its service is started, the
first call to fleetctl must be retried a few times.  Calls to block
until the etcd cluster is healthy have been added to help reduce
this time between SSHing and fleetctl working.