journal: add Entry type and field names
Wrapping the entry map cleans things up a little and adds a place for
utility methods to go, such as this first one: a timestamp parser.

Define all common field names as constants, using these help catch typos
that are easy to make when using bare strings as map keys.