Add SSH helper method to cluster
This will be used to replace all uses of cluster.Masters[0].SSH() that
litter this code base and end up printing a lot of junk out to the test
results. I get lots of questions about this. The solution outlined in
https://github.com/coreos/mantle/pull/426 is certainly better for
dealing with stderr, but we can unify with that implementation later.