kolet: fix everything
When kolet was converted to use cobra subcommands in c56732e4 the final
`os.Exit(2)` fallback never served its purpose because `cli.Execute()`
always exits the program. This was made worse by the fact that cobra
does not consider running a command without a `Run` method to be an
error, simply printing the `Short` description and returning. When test
registration was working this wasn't noticed because cobra would raise
an error if you did not provide a valid subcommand. In commit 3a073663
test registration in kolet broke because the import line was not
updated. So for the past few months all kolet based tests merely checked
that kolet could execute and print the `Short` description from the run
command. This was further made impossible to notice because the output
of kolet is never reported unless it exits with a non-zero value. We
probably could have noticed the kolet breakage in one case but fe34b448
worked around the issue by removing the kolet usage instead of finding
the root cause.

On the up side, we never really lost much test coverage because all of
the remaining kolet tests came from coretest which we still use in the
old AMI test script that we are still working on replacing.