*/aws: Delete S3 object after creating snapshot
Split the part of CreateSnapshot() that looks for an existing snapshot
or task into a separate FindSnapshot(), and have plume and ore upload to
S3 only if they need to initiate creating a snapshot. That allows us to
delete the S3 object after we're done with it, without causing
unnecessary uploads when rerun.