kola: restore pause before cluster tear down
Honestly I'm not sure how important this is and would be better
performed as some sort of flush mechanism in the journal recorder but
implementing flush would be a bit complicated so this stays for now.
It was broken by the conversion from returning errors to using
Error/Fatal methods rendering this `err` value always nil.