cmd/plume: List Azure environments in spec
Previously the list of environments came from the Azure profile file,
which is not version-controlled alongside plume.

Also split the container list into one environment-independent container
name (the same name is used in each environment) and a per-environment
list of additional containers. The first container in the list was special
(it's the container that's actually used to back the OS image) and this
makes it explicit.