platform/aws: automatically create SGs
If they don't exist, create them rather than failing when running
instances.

This was tested in VPC-enabled regions where it'll work fine. In non-vpc
regions (ec2-classic regions like us-east-1 on many accounts), things
are a little more complicated, but this doesn't make it worse than
before.

I also considered adding this to initialize or making it optional based
on whether an SG parameter was provided or not, but in reality I think
this behavior is basically always what we actually want so I don't see
the need to complicate it with more options currently.