cmd/plume: Wait for existing GCE image to finish creating
We can't assume that any GCE image we can find is complete; it may still
be creating from a terminated previous plume run. Wait for completion in
that case.

Because we need to be able to resume images named with the previous day's
datestamp, this logic is going into plume rather than CreateImage.
Therefore, unlike with AWS, "ore gcloud {create-image,upload}" do not
inherit the ability to resume existing images. When run without --force,
they will fail on a previously-created image rather than waiting for it
to complete.