cmd/plume: Allow pre-releasing each platform independently
No platform has pre-release dependencies on any other, so allow selecting
a subset of platforms to pre-release. This is intended to support
pre-releasing platforms separately in Jenkins. Similar functionality is
not extended to "plume release" because (a) "release" is supposed to be
relatively fast, and (b) we don't want to risk performing a partial
release. "plume release" will catch missing artifacts on each platform,
so a partial pre-release won't slip by undetected.