platform/qemu: Add method and subcommand to set serial as primary console
kola can't check for initramfs error output if that output doesn't go to
the serial port. Add a kola subcommand to copy a Container Linux image.bin
and modify the OEM grub.cfg in the copy. Make this a separate operation
for now (e.g. to be run in Jenkins) because a) it's not mandatory, b) it
requires kpartx, and c) it could leak an image file if it was integrated
into "kola run" and kola was killed.