kola: By default, put results in a subdirectory of _kola_temp
Put test results in a subdirectory of _kola_temp named after the
platform, date, and pid.  Maintain a <platform>-latest symlink in
_kola_temp that points to that subdirectory.  This prevents simultaneous
test runs from clobbering each other and makes it trivial to preserve
history across runs.

We don't maintain a platform-independent "latest" symlink because the
platforms that test as root (qemu) would fight over permissions with the
platforms that don't.

If --output-dir is specified, continue using it directly.