kola/tests/network-listener: check for listeners with netstat
The lsof checker in use before had occational false positives on UDP
since it was unable to distinguish listening from requesting.
Specifically, timesyncd appeared a couple times even though it was only
making a DNS request.

This switches to netstat which, as best I can tell, has better filtering
for 'listening and udp'.

This also changes ports to be numeric simply because I find that easier
to read.

Functionally, this should be the same as before.