tests/kubernetes: add kubelet var-log mount test
This is a regression test promised in https://github.com/coreos/bugs/issues/2016

This is the stupid-simpleist way of testing that specific regression.

This can be replaced with another test that exercises kubelet-wrapper
additionally, but retains this bindmount, in the future since this test
does run rather slowly for what little work it's doing.