kola/tests/ignition: update ext4checkexisting for current Ignition
The current test validates mkfs.ext4 -p, but we no longer carry the
e2fsprogs patch for that option. Ignition now detects existing
filesystems itself via blkid, and that functionality should be tested
by its test suite, but it's still useful to have one integration test
here.

Testing the non-force case with pre-2.1 configs would produce a boot
failure, which is difficult to check. Use wipeFilesystem=false and a 2.1
config to test that Ignition doesn't reformat a filesystem with the
correct label.