kola/tests: test enabling a torcx systemd unit
This introduces a new torcx package even though this test could also fit
under the existing docker package. The primary reason is the intent of
this test is for a torcx feature.

It's possible in the long term that this testing should live under the
torcx project itself, but for now I think the testing living in kola is
pragmatic.