kola/tests/rkt: add systemd+rkt integration tests
This test is meant to catch a regression in v1492.0.0 whereby `rkt run`
would fail under a systemd unit.

Since Container Linux is the point where a pair of (rkt, systemd) are
run together, and the default rkt `stage1-coreos` interacts with
systemd, it makes sense to test this integration here, not just in rkt's
repo.

The tests introduced here attempt to cover the user-observable details
of the integration between rkt/systemd.