kola/etcd: add etcd-member and etcdctl tests
This commit adds a pair of etcd-member tests since our coverage for it
was so low.

The first, v2 backup/restore, tests doing a backup / restore of a
single-node etcd cluster.

The second, etcdctlv3, covers basic operations of the v3 cli.

Both are implemented against size 1 etcd clusters for simplicity.