platform/base: Render CLCs with platform
This updates the BaseCluster to know what ctplatform it's for, and for
each of those platforms render Container Linux configs appropriately.

Platforms CT does not have specific support for pass in the empty
string, which results in them working fine for any CLCs that do not use
special variables.

These platforms can still use CLCs for userdata as long as they don't
reference special platform-specific variables.