kola: add torcx manifest option
This adds a flag so that a user running kola may pass in a torcx
manifest.
It's expected that this manifest will be provided at the very least on
the automated jenkins runs of kola.
Notably, this manifest is not sourced from a remote location. This is
important as it allows these tests to run against a manifest before it
is signed and published (else, what's the point of testing it anyways if
it's too late to catch problems?).

This also works reasonably well for developers who wish to test torcx
packages locally.
It would probably be more convenient if kola, when in an SDK,
automatically used the "well known" latest manifest location from the
build directory.
That was omitted for simplicity, but may be added in the future.

The manifest was plumbed through and exposed to tests in the simplest
way I could figure out how to.

In some future magical version of kola, this option may also result in
kola hosting the torcx pkgs referenced by the manifest within the
testcluster and rewriting the URLs so tests can correctly request them.
That's also omitted since that would be a much more significant effort.