kola/tests/docker: add torcx manifest pkg tests
This downloads and tests all docker packages in the referenced torcx
manifest.

It is somewhat wasteful in that the default torcx package is tested
twice.
That can be skipped as a future optimization.

It's intended this will be run as part of the the kola tests run after
pre-release by having jenkins download and pass in an appropriate
manifest.