kola/tests/docker: Remove /var/lib/docker before changing versions
Docker 17.* attempts to use an existing graphdriver by default, but
the test expects its fallback "overlay2" driver.  When a system is
booted on 1.12, it will be using "overlay", so the test fails when
it reboots into 17.06 and still has "overlay".

This wipes out the /var/lib/docker directory before starting the
new version after reboot, so Docker 17.06 will start using the
"overlay2" driver.  This effectively changes the test to not be
concerned about testing in-place Docker upgrades, rather it is just
reusing the same system to test different versions for efficiency.