platform/conf: add IsEmpty function to Conf
The OCI platform does not support cloudconfig. As such, the cluster
errors out if a non-ignition config is detected. However, empty
configurations should be valid. `IsIgnition` correctly returns
`false` when the Conf does not contain an Ignition config, so this
adds a new function `IsEmpty` which will return true if no
configuration data is present.