cmd/ore: add upload oci command
Adds an `upload` command for the OCI platform. This only
uploads the image to objectstorage, it does not import the image.
Currently the `go-bmcs-sdk` does not support importing the image
from anything other than a running instance.

Until the support is added you'll need to manually create the image
via the `oci` command-line tool or the web UX.