cmd/kola: warn in PS1 if exiting shell will destroy spawned machine
When running instances spawned with "kola spawn" alongside instances
started in other ways (including "kola spawn -t"), it's easy to
accidentally terminate a machine by exiting a shell. To avoid this,
when starting a shell which is bound to the lifetime of its instance,
add "[bound]" to its shell prompt. Ideally this marker would only
appear in the affected shell (and not shells from other SSH connections)
but there doesn't seem to be a good way to do that, so it will appear
in every shell on that machine.