kola/tests: disable tests on Azure
The Azure SDK requires a hostname to be set via the API during machine
creation and then overwrites the hostname post-boot via wa-agent.

Update the metadata/contents test to only validate
COREOS_AZURE_IPV4_DYNAMIC as the machines being created are not behind a
load balancer.

The disabled tests on Azure are:
 - coreos.ignition.v1.sethostname
 - coreos.ignition.v2.sethostname
 - coreos.install.cloudinit
 - linux.nfs.v3
 - linux.nfs.v4
 - systemd.journal.remote