coreos-virt-install: New script
This is *way* simpler than ImageFactory, and much more widely used.
In fact it's documented in the Red Hat Customer Portal, etc.
We avoid the whole ICICLE stuff which is really irrelevant to us.

And most importantly, it Just Works in an unprivileged container
provisioned with `--device /dev/kvm`, since we use `qemu:///session`.
(I tried to pass that though ImageFactory/oz but wasn't successful).
This will allow us to build CoreOS inside containers, as we should.