build: Clean up/doc dependencies, stop building rpm-ostree
First, let's just use the latest rpm-ostree release which has
the features we need; building it (and installing the builddeps + removing them)
was really really slow.

Next, move the packages we're installing into variables that
are documented (so we know why they're there) and do them as much as possible
in one group.