Chown everything to builder, stop using ACLs
I thought I was being clever using ACLs, it *mostly* worked
but the problem is that some of the objects aren't world readable,
and due to the way [effective ACLs](https://unix.stackexchange.com/questions/152477/how-does-acl-calculate-the-effective-permissions-on-a-file)
work, those objects weren't readable to `builder`.

This approach of doing the chown after treecompose matches what
FAHC does too.