New command: run
This is a copy of my
https://github.com/cgwalters/playground/blob/master/coreos-run-qemu
script.  Which I know overlaps a lot with `kola spawn`, but as
is noted in the header:

 - Uses `-snapshot` by default to avoid writing to the disk
 - Contains a default  Ignition config that auto-logins on the console

Also, this runs as non-root.

The main change I made from the playground one is small but important
one:

`coreos-assembler run`

finds the latest disk image and runs it this way.  I find this *very*
convenient.