build: Add Rust
So far we have lots of shell script, some Python, Go from mantle...
clearly, we need to add Rust!

More seriously so...I like the language, I am not excited about
writing more dynamically typed and slow shell script/Python.

There's a definite argument for Go, since we could share code
with Mantle.  I personally like Rust a lot though...but
this is an RFE to see how people feel.

My vision for this was that we could start implementing some of
the VM installation work in it, along with defining Serde
structs for the metadata, etc.