virt-install: Disable networking
We don't want Anaconda to find a network device and do DHCP,
and try to encode anything about the build environment.

Further, we want all of the content to come from the container;
nothing potentially from the Internet.  Downloading content in `%post`
is a no-no.

Since we currently inject convent via 9p, disable networking
in the installer.

This will break RHCOS builds, but we'll adapt them to use this.
Right now it's tagged to `:alpha`.