oscontainer: New command
For Red Hat CoreOS, we plan to ship the OSTree inside a container,
or "oscontainer" for short.

This script offers two subcommands: `extract` and `build`.
Now, an important point to understand is that these tools are
expected to run from inside an existing container.  Since
overlayfs-on-overlayfs is not supported, we use a separate
`--workdir` which translates into `podman --root`.  Hence
these tools have nothing to do with any container storage on the
host system.