cmd-run: Add coreos-assembler-quiet-tty.service
Right now, we're running at the default log level, which is DEBUG (7).
The text gets interspersed with user input/program output on the tty.
Bump the default to ERROR (3).

This uses a systemd service hack instead of a `sysctl.d` drop-in. The
reason is that Ignition relabeling right now happens too late in the
process for `systemd-sysctl.service`, so it trips on SELinux issues.
Looking to fix that soon.