cmd-[build|fetch]: Drop set -x
The scripts are becoming more involved, and all the output we get from
`-x` mostly distracts from the meaningful output. Let's just drop those
and we can always add back some strategic `echo` statements if we feel
like any are missing.

(Note we still leave it in `cmd-clean` for now; it's much shorter and
it's an easy way to see what's getting nuked.)