WIP: Clean /var after Anaconda
This way we are truly starting from a clean slate.  It means
we know we won't have a `random-seed` for example.  OSTree is
designed to support this but unless we use it *by default*
we aren't really testing it.

This actually now breaks things since currently the creation of
the `core` user is in `image.ks`, which we need to fix:
https://github.com/coreos/coreos-metadata/pull/90#discussion_r202438581