cmd-build: Preparatory work for supporting builds without refs
If one is doing builds via oscontainer, it doesn't make sense
to use refs as well.  Rework the build process so that most
of it doesn't require specifying a ref.

Today however, Anaconda does require a ref, but we're heading
down the road of not using Anaconda.

Another thing this takes a step towards is supporting a mode
where the ref is only written "for real" when the image build
is complete - we want to support people truly binding them
together.