Consistently set tmp/treecompose.changed
This will allow e.g. an orchestrating Jenkins job to look for
the file afterwards and only trigger a `build` if something changed.
This is better than doing it in one job since then the secondary
build job will only have actual builds.

This also fixes a subtle bug in that we weren't clearing the stamp
file before doing a build, so we could spuriously think something
did change.