build: Move "no changes" logic earlier
When I went to deploy coreos-assembler in Jenkins, I hit a
crash in the case where nothing had changed.  The reason is
that in this case the `tmp/` dir isn't saved across builds, and
we were relying on the previous build's treecompose metadata
being cached there.

Fix this by moving the detection of the "no input changes"
case to follow immediately after the treecompose "no changes" logic.

Parsing of the previous build image data, and computing the kickstart
checksum also happen before treecompose now, consistently alongside
the places where we parse the ostree data out of the previous build.

Now, there's another corner case here - in the orchestrated case,
how do we handle the case of *just* the `image.ks` changing?  What
I chose to do there is to pull the previous build's `meta.json`.