build: chown repo in privileged mode
This is a regression from the unprivileged work.  Previously
we had two repos, and only the `repo-build` was owned by root.
Now that we have one repo, in the privileged path `rpm-ostree` still
runs as uid `0` by default, so we need to `chown` the repo to
ensure that e.g. `ostree prune` (run unprivileged) can operate
on it.