tests/check.sh: disable ShellCheck if older than 0.4.0
Split out from #270

TL;DR - ShellCheck is disabled unless version 0.4.0 or newer is
installed.

The ShellCheck test makes use of the `--external-sources` option, so
we can load in external libraries during the checks.  But that
option is only available in version 0.4.0 and newer.  Older versions
of ShellCheck will barf on the `shellcheck source=src/cmdlib.sh`
directives and trying to tell ShellCheck to exclude the raised error
was not successful.