Use stdlib for date parsing
pytz and dateutil are not available for all bases we want to build
with. Instead of trying to create new packages and get them approved
this change moves to using the python standard library. While not
as pretty, the same or similar enough functionality is provided.

NOTE: All dates are treated as UTC.

Signed-off-by: Steve Milner <smilner@redhat.com>