.travis.yml: switch to use native tests/check.sh script
This overhauls the Travis CI job to use the native `tests/check.sh`
script.  This required the Travis enviroment to have the latest
version of `ShellCheck` manually installed, but gives us parity with
the experience of using `check.sh` on a F29 host.

I left some lines to print the version of `python3` and `ShellCheck`
just for the occasional debug need.