tests/check.sh: change to check all executable scripts
Previously, we were only checking the files in the `src` directory and
erroneously leaving out the scripts like `build.sh`,
`coreos-assembler`, and `bottlecap`. This changes the behavior of
`check.sh` to build a list of files to check via a whitelist'ed set of
directories and then runs the appropriate syntax checker against them.