build.sh: switch commitmeta_to_json to python2
Because there is no `gobject-introspection` module available to the
Python 3.6 stack from the SCL, we have to use Python 2 in the EL7
container case.  This will change the shebang on the script to use
`/usr/bin/python2` (for EL7 only), which avoids the SCL and
provides the ability to use `import gi`.