Support image.yaml, move default kickstart here
We want to have *some* configuration, but it shouldn't be kickstart
for multiple reasons (we want to possibly remove Anaconda).  For
now while we continue using Anaconda on the builder side, let's add
support for a very basic YAML config file which basically
allows us to say e.g.:

`size: 8`

And that's all.  This way people forking the config don't need
to also fork all of the implementaton details like removing
the random seed etc.

To ensure we handle version skew (new assembler, old config)
for the existing community coreos-assembler users, we continue
to support using an `image.ks` - for now.