kickstart: enable prjquota for quota enablement
This enables platforms like kubernetes to limit the storage of
containers so malicious actors in a multi-tenant environment
can't eat up a ton of storage.

We can't do this in the fstab because the option needs to be enabled
on the initial mount (during initramfs) and our initramfs files are
generated during ostree generation so we'll use rootflags kernel arg
for now.

https://bugzilla.redhat.com/show_bug.cgi?id=1658386