src/cmd-build: Add `ostree` target
I'm often just iterating on the OSTree content itself and don't require
new images. In such cases, it's a waste of time to wait for images to be
generated when all I want is the treecompose. But using
`coreos-assembler build` is still more convenient than dropping down to
`rpm-ostree compose tree`, especially with the virtualization wrapper.

Change the concept of `IMAGETYPES` to `TARGETS` and add a new `ostree`
target for which we only generate a commit to the OSTree repo. This is a
natural extension of #302. Crucially though, we still create a build
directory under `builds/` with metadata about the built commit. The only
practical difference is that there are no image files and no subkeys
under `images` in `meta.json`.

(I didn't bother breaking the idempotency coupling here wrt the
kickstart since (1) it's going away, and (2) this is really just for the
local dev case where you're iterating on the tools, like rpm-ostree, or
the content, like the treefile.)