sdk/enter: make compatible with Silverblue
On Silverblue, `$HOME` is in `/var/home`, so that throws off code that
assumes otherwise. The rest of the SDK assumes `/home` is where the user
home directory will be inside the chroot, so let's keep that assumption
going by building the path to `/home` here too rather than passing
through the original path.