runvm: support setting environment for supermin appliances
When running some commands in unprivileged mode, such as podman or
buildah, the supermin environment needs extra credentials. This change
allows for extra files to be added to `tmp/supermin` and
`tmp/supermin/supermin.env`. `runvm` will expose `tmp/supermin` and to
the appliance and set the environment from `supermin.env`.