Support not specifying a ref in manifest
For Red Hat Enterprise Linux CoreOS, we will be shipping updates
via "oscontainers"; see the `oscontainer` command history for
some more details.

Support not specifying a ref in the manifest.  The semantics
result in the ref being deleted in the disk image.

Today the RHCOS build system adds a bit in the kickstart to do
this, but we want to get away from supporting external kickstarts.

This patch is my 3rd attempt at implementing this; it turns out
to be by far easiest to override the manifest and add the ref temporarily.
Anaconda today requires a ref and not a commit.  And if we didn't
have a ref in the local repo, then the prune logic goes wrong, and
it also breaks `automatic_version_prefix` etc.

So we have a `tmpref-${name}` ref in the repo, and just strip
it from the generated disk image.