build: ensure git submodules have been initialized
If the developer hasn't initialized submodules before attempting to
create the coreos-assembler container image, they'll be met with an
error similar to the following:

    error: submodules not initialized. Run: git submodule update --init

Adding this to the Dockerfile ensures that this command has been run. It
is safe to run multiple times, so this shouldn't introduce problems for
those who initialized submodules before the container build process.