buildextend: create initramfs and kernel artifacts
The installer ISO generation already pulls the initramfs and kernel
from the ostree repo so this just piggybacks off that and copies the
artifacts to the build directory.  Having the initramfs and kernel is
useful for tools like virt-install and OZ to look at a tree and boot
the kernel with fine grain control over the kernel command line.

Do not compress the kernel or initramfs.img because they can be used
directly by URL in iPXE.