build: Save raw commit object in builddir
Save the raw commit object in the builddir;
it can be useful to provide the previous commit during
builds, and this avoids a need for the full OSTree repo.

Particularly for RHEL CoreOS where we ship via oscontainer, this
will help us avoid needing to fetch the previous oscontainer, when
all we need is the previous commit object.