virt-install: Exit with error if port is in use
When I installed the OpenShift extension for VS Code that started
spawning a JVM in the background which for probably terrible
reasons listens to port `8000` by default.  We should also
do dynamic ports at some point.

Then it also turned out that Python's raw thread API has
the rather unfortunate semantics that an exception thrown in a
thread just prints to stderr and continues.

I think we really want is to use the futures suggestion from
https://stackoverflow.com/questions/2829329/catch-a-threads-exception-in-the-caller-thread-in-python
but I'm not feeling it's worth trying to figure out how to rearchitect
this code to use futures right now.