cmd-run: Drop --persist and add --persist-to-img
Using the current `--persist` is rarely a good idea. Let's instead
always point qemu to an image which is backed by the target file. A new
`--persist-to-img` flag is then added to select whether that image is
temporary or re-used between invocations.

This is useful when we want to be able to hack on a VM that has a bit
more persistence, but not necessarily go the full `virt-install` path
(or if you can't use libvirt for whatever reason). Basically make using
`cosa run` as a glorified `qemu-kvm` easier.