cmd-build: Consistently give ref or commit to Anaconda
Right now, we were passing `${commit}` to Anaconda in the
buildextend-metal case, but that meant that even if the ref wasn't
temporary, we would end up with a checksum refspec.

We were also passing `${ref:-${commit}}` for the qcow2, which would
always evaluate to `${ref}` since even if the manifest defines no ref,
we create the temporary one.

Instead, always give `${ref}` to `run_virtinstall` and add logic there
to "dereference" the ref only if it's temporary. This is the case for
RHCOS, though it doesn't actually matter much for that one since it
"prepivots" images anyway.

Though this does fix the case of FCOS bare metal images having a
checksum refspec.

This does regress the `buildextend-metal` case when targeting a build
that's not the latest by feeding it the latest ref instead of a specific
commit. This will be fixed in an upcoming patch.