Detect rhcos for Ignition spec 2, add start of 3to2 converter
It would be really extremely convenient for me if I could at least
`cosa run -d /path/to/rhcos.qcow2` when using cosa git master.

This prototypes out the start of a `ign-3to2` helper which
will also be extremely useful for people working on OpenShift/RHCOS
today.  If this works out we can elevate it to a toplevel command.

I reworked the `abspath` bits as it turns out my use of `git annex`
didn't work with this as the symlink ended up being canonicalized
to the object name.