Also keep around /var/home by default
A lot of backstory here, but essentially, we want Ignition to be able to
add users in RHCOS. This requires `/var/home` existing. To do this,
we either need to run systemd-tmpfiles, like FCOS, or just leave it
hanging around after Anaconda.

Since we're already doing the latter for `/var/lib` and `/var/log`, this
patch just extends it for `/var/home`. It's just easier to do this for
now rather than more involved fixes. This will be fixed the proper way
once RHCOS is bumped to spec 3 and inherits all the work that happened
in FCOS to support it.

Some more information about this in:
https://github.com/coreos/ignition-dracut/pull/79