cmdlib.sh: Use git timestamp for OSTree layers
That way the commit checksum itself is completely stable when
recomposing on the same git commit (modulo a dirty directory, but then
the checksum will change anyway).

Doesn't really matter for now (except for the good feelsies of always
getting the same commit hash between invocations), though it will if we
want to start recording more information about constituent layers in the
compose.