cmdlib.sh: Don't `chown` in supermin path
Only root can `chown` things. Though we're root in the supermin
appliance, files we create on the 9p mount are all created with our
uid, of course. Doing `chown -R` like this worked before because
`chown`ing to yourself isn't an error.

But now with the Anaconda removal work, we use the supermin path even
when privileged. Except we don't have a `cache.qcow2`, so our root-owned
cached (since rpm-ostree is still run as real root) was showing through
the 9p mount, and `chown` was no longer a no-op and would of course
fail.