src/create_disk: use more reasonable sizes
Use 384M for /boot since it may have up to three (two?) kernels +
initramfs's. With Container Linux we could get away with 128M since our
kernels and initramfs's were smaller, but we'd likely hit issues with
the fedora ones.

Dramatically shrink BIOS-BOOT. 1M should be plenty. The grub docs refer
to 1M as being enough[1].

Shrink the ESP by 1M. We hardly put anything in it. Could probably stand
to be smaller but I'm not sure the ramifications right now. All that's
in it is the shim, grub efi program, and a tiny stub config.

[1] https://www.gnu.org/software/grub/manual/grub/html_node/BIOS-installation.html#BIOS-installation