gcp: Always --force
Our model is built on "immutable" version numbers, however it
can happen that e.g. a GCP build completes but a later stage fails.
If that happens, we may reassign the version number, and the correct
thing is to overwrite the build.

So always pass `--force`, and rely on the higher level pipeline
logic not overwriting builds that were successful.