cmdlib: Add helpers for fetching `${basearch}`
Add some helper functions for retrieving the dnf value of `${basearch}`.

Note there are many places over the codebase where we use `uname -m` or
`arch`, but we semantically mean `${basearch}`. I didn't bother
replacing them because (1) for the architectures we care about, it
actually makes no difference (if e.g. one day we want to support 32-bit
ARM or x86, then it will) and (2) given the above, the refactor risk
didn't seem worth it.