cmd-build: Add `${basearch}` to commit metadata
Mark the OSTree commit with the actual `${basearch}` we used to assemble
it. The major user for this will be Zincati. Putting it in the commit
metadata means this will then show up in e.g. `rpm-ostree status
--json`, which Zincati already executes today.

We could also do this directly in rpm-ostree. Though really, after
digging into this quite a bit, I lean towards making this a CoreOS-ism
for the time being. There aren't many software out there that care about
the specific `basearch` string; it's something that mostly only matters
to yum/dnf/rpm-ostree (hence why it's not really exposed in any easy
way).

As one can guess, any portable software that needs arch-specific
behaviours would use e.g. compiler flags or `uname(2)`, rather than any
kind of distro-specific concept.

I also didn't make this part of `/etc/os-release` because if Fedora does
end up shipping it there as well eventually, we'll have to carry *two*
variants of the same tag. OTOH, keeping it in the commit metadata forces
the few classes of software that care about this to be specific to
rpm-ostree systems, over which we likely have more control.