cmd-build: Check if previous OSTree tarfile exists before copying
If we initialized with `buildprep`, we won't have `ostree-commit.tar`.
If `cosa build` then detects no changes to the OSTree, but only to the
image it will try to avoid tarring by copying the previous build's
OSTree tarfile, which will fail. Just add a check for the file before
enabling that optimization.