Add support for lockfile overrides
This is needed in FCOS, where we'll have a shared base lockfile, and a
curated overrides file. For more info, see the stream tooling doc[1].

For now, we just keep separate overrides per basearch to keep things
simple (instead of having a single override as per the stream tooling
doc). We can add more enhancements in rpm-ostree to enable this, but
let's get the basics working first.

[1] https://github.com/coreos/fedora-coreos-tracker/blob/master/stream-tooling.md