grub.cfg: Fix ignition.firstboot semantics
A very important piece we lost moving towards a static GRUB2 config is
the `rd.neednet=1 ip=dhcp` args which normally comes from
ignition-dracut[1].

It also included some subtle semantics which interact with
coreos-installer through the `ignition.firstboot` file so that
networking args are carried over.

Not sure if there were plans to deal with this differently, though we
need something for right now at least for FCOS preview. (And we'll want
to drop `02_ignition_firstboot` from ignition-dracut, but it's not a
hard requirement).

[1] https://github.com/coreos/ignition-dracut/blob/8551bc7fd89a80200d94d009ad3d17131fc11dcf/grub/02_ignition_firstboot