cmd-build: Add --parent
We want to have full control over the parent of an OSTree commit so it
can be driven at a higher level. IOW, cosa as used in a CI/CD setup is
not in a position to know what the parent of the commit should be. So
either it should default to *no parent*, or accept an override for a
specific parent at build time.

This will be used by FCOS at least for the time being. See:
https://github.com/coreos/coreos-assembler/issues/159#issuecomment-512594688