kola/tests/crio: change how pod names and images are handled
The most recent version of `cri-o` is preventing new pods from being
created that have the same name, so this changes how the names are
generated for the pods.

Additionally, we were re-creating the container image each time the
pod was created, so change the logic to test for the container image
first before we try to create it.