cmd-run: Make --srv work again via dependency hacking
This is an awful hack-around
https://github.com/coreos/fedora-coreos-tracker/issues/223

But I use `--srv` for developing a lot as it's an easy
way to get data into the VM (e.g. test RPMs), so let's
make it work so I and others can use it to develop FCOS
and even test fixes to things like this.