buildextend-metal: Add boot partition sizes to estimate
Similar to what we were doing with the virt-install/Anaconda path,
which adds the size of the boot partition (1GB) to the estimate.

`cosa buildextend-metal` was hitting `ENOSPC` for me building
Silverblue.

(In the future when this isn't shell script we should obviously
 only have the ~512 MB in one place, but for now this is better)

While I have the patient open, change the metadata percent estimate
default to 20 to match what we were passing anyways.