build.sh: Use cmdlib.sh from src/ dir
For building, we were putting `cmdlib.sh` in the root of the container
build dir. But I want to be able to use `build.sh` from a git checkout,
so rebuild that same layout.

Also drop the direct reference to `/root/containerbuild` for the same
reason.

And also add the motivation for why the `Dockerfile` calls `build.sh` in
multiple steps like this.