ci: Add testing using Jenkins pipeline
Start running tests in CentOS CI. We build it, run the unit tests, and
in parallel run through a simplified version of the FCOS pipeline
including running `kola` on the built image.

This is just a start of course. In the future, we could e.g. test each
command more deeply, such as lockfile semantics, overrides, etc...

This uses the `projectatomic-ci` namespace for now, where we have
resources and permissions set up. Once we're willing to commit more to
this workflow, we can move to a `coreos-ci` namespace (and let
`projectatomic-ci` die peacefully).