create_disk: Set up SELinux labeling fully
Booting current FCOS, one can see that stuff under
e.g. `/sysroot/ostree/` is mostly `unlabeled_t`.

Also, there's junk in `/sysroot` that isn't necessary; I think
when I was writing `ostree admin init-fs` I was thinking of dracut
requirements at the time that the target root (before ostree pivot)
had e.g. `/dev` - that's no longer true (if it ever was).

For each mount point, add an initial hardcoded label; the physical
root is `root_t` for example.  This ensures that at least anything
created under that will have *a* label, and usually via
path transitions, the correct one.

Combining the above, we inline the bits of `init-fs` we need, namely
creating the repo and `deploy` directory.  We add the `usr_t`
label to the ostree repo, and let SELinux policy handle the rest.

Then the result is that the full disk we create should have
SELinux labels on all files.

This is just a good fix in general, but I'm writing this specifically
as prep for having `/var/home` pre-created to aid RHCOS, which has
an older SELinux policy that doesn't allow Ignition to operate
on unlabeled data.