build: Readd support for save-var-subdirs-for-selabel-workaround
Currently the SELinux policy in RHCOS doesn't allow systemd
to operate on `unlabeled_t`.  General agreement was
that the real fix is to do some "bootstrap" labeling in
Ignition, but the problem is the kernel doesn't currently
allow it.

For more information, see:
https://github.com/coreos/ignition-dracut/pull/79#issuecomment-488446949

For now, we just carry forward this (opt-in) workaround which already
exists in the `virt-install` path.

Closes: https://github.com/coreos/coreos-assembler/issues/581