src/cmd-aws-replicate: Add command to replicate AWS AMI to regions
Adds a new command `aws-replicate` which will replicate the AMI of a
given build to a list of regions. It utilizes the `ore aws copy-image`
command to do the actual copying, searches for the initial AMI from the
buildmeta, skips any given regions that already exist in buildmeta, and
finally updates the buildmeta with the new values.