platform/api/aws/images: return snapshotID from CopyImage
Snapshots are created when copying images to new regions, these extra
resources cost money and should be communicated by the API. Add
additional code in `plume` to throw out said snapshot IDs as they are
not published to end-users.

The end result of this PR from the perspective of command interactions
is that `ore aws copy-image` will have the snapshot ID included in the
JSON output.