cmd-buildupload: check that files exist
Currently `buildupload` doesn't actually validate that the images
described in the `meta.json` exist in the current build directory before
attempting to upload them. This can be problematic in the case where a
user has performed a `buildprep` to pull down an existing builds
`meta.json` and is now attempting to re-upload it after adding new cloud
images / artifacts.