README.md: Document cosa via unprivileged podman as primary
The advantages of this are *compelling* - among others, we don't
need the ACL dance, and it emphasizes how one can do builds as
non-root.

The doc has a lot of `sudo podman` which I've excised.

A downside of this is it's less documented how to get things
to work with Docker; I didn't try that, but it should be really
easy.