Add a Dockerfile.buildroot
We've talked before about how cosa isn't useful for people
developing on e.g. rpm-ostree; this is a primary reason I personally
use cosa via installing it inside my personal toolbox container.

This PR implements something talked about before; we add a
"buildroot" container (e.g. `quay.io/coreos-assembler/coreos-buildroot:latest`)
that is basically `yum builddep` for the parts we care about.

The main reason I want to do this right now is I am working
on using cosa in rpm-ostree's CI, and having an up-to-date "buildroot"
container with the dependencies pre-installed will help a lot with that.

I ran into issues trying to use cosa as a base and do the `yum builddep`
dynamically because of the `USER builder` directive; Prow isn't expecting that.

Another reason to do this is that in the future I'd like to
add first-class support to cosa for e.g.
`ln -s /path/to/ostree overrides/src`
and cosa would take care of building a layer (or RPM?) from that
source.