cmd-build: Split out qemu building into separate command
For FCOS signing, we need a way to build just the OSTree, then have it
signed, and then build the image artifacts with the signatures embedded.
To do this, we need to decouple qemu image building from `cosa build`.

This patch does this by making `cmd-buildextend-qemu` a symlink to
`cmd-buildextend-metal` and enhancing the latter to handle both cases.
A big advantage of this is that base image building is now located in a
single location.

We then extend `cosa build` to just take any arbitrary `TARGET`
arguments which become semantically equivalent to calling `cosa
buildextend-$TARGET`, where `TARGET` is one of `ostree`, `qemu`,
`metal`. (And of course, we default to building the qemu image to not
break compatibility).