buildextend-live: allow patching live ISO to embed Ignition config
Pad the live ISO initrd to leave space for an external tool to patch in
a cpio.gz archive with a user Ignition config.  After generating the
ISO, calculate the offset of that padding and embed the offset/length in
a header at the end of the ISO 9660 System Area.