run: Support specifying both build ID and image type
I want to make it easier to run a metal image.  First we add
`-b/--buildid` to specify a build ID, which is a generic improvement.

Next, add a `-I/--imgtype` argument which is most useful for e.g.
`cosa run -I metal`.  This automatically turns on injection into
the `/boot` partition.