s390x: do not write first boot args to BLS file
The first boot arguments are just ephemeral. If we add them to BLS file,
then we have to remove them in ignition-firstboot-complete.service
during first boot, which would complicate it. Without adding them to BLS
file, an invoke of zipl without any parameters (assuming zipl.conf and
BLS files are present) is enough.