cmd-buildextend-metal: support for rootfs in LUKS containers.
This will install the rootfs in a LUKS container using a null cipher. The
actual encryption of the LUKS container will handled by a new dracut
module, tentatively named "ignition-dracut-reecrypt."

By default, the rootfs is not installed into a LUKS container. To enable
rootfs in LUKS, add the following to the recipe image.yaml:
    luks_rootfs: "yes"

This implements the agreed-to-design in:
https://github.com/openshift/enhancements/blob/master/enhancements/automated-policy-based-disencryption.cmd