cmd-build: rename OSTree commit tarball
To be more consistent with the other artifacts, rename the OSTree commit
tarball to include the OS name, build ID, and base architecture.

Note the previous patch added the `path` of the tarball in `meta.json`,
so this patch also teaches cosa to read the tarball name from there and
fallback on the legacy `ostree-commit.tar` name.