buildupload: Drop --freshen argument
We added this when the OpenShift installer was *directly* pulling
from the RHCOS build stream.  They wanted to be resilient to
"update denial attacks".  Since then, the installer pins RHCOS
and actually no customer-facing tool talks to the RHCOS cosa build
stream.

So we don't need to worry about "freshening" it.  I also
dropped the call to `--freshen` in the RHCOS pipeline a while
ago.

Prep for further work in this file around build cancellation.