Remove support for "legacy" build schema
The new build layout includes architecture-awareness.
It was really nice that we supported both for a while, but
RHCOS for 4.3 has switched to the new layout.  There's no
reason to keep around the legacy build layout code, so remove
it for at least three minutes of warm fuzzy "code deletion high".

Also prep for work around failed builds handling.