cmd-run: Add `--ssh` option
With `--ssh`, one can connect to the guest via SSH instead of serial
console. The main advantage of this is a nicer terminal (as well as
interacting with the VM in the same way that most end-users would).

To make sure we always GC the VM, we use `prctl(PR_SET_PDEATHSIG)` on
`qemu` and then simply `exec` into `ssh`.  The disadvantage of this is
that rebooting breaks the VM (though one could just use `--ssh-config`
instead).

And yes, there's definitely an overlap here between `cosa run` and `kola
spawn`, and this muddies things a bit more. Right now, I think more of
`cosa run` as a lightweight runner (specific purpose, dead simple
process tree, doesn't write out any files) whereas `kola spawn` is
heavier weight. Not to say that we can't fold those features in `kola
spawn` though (we definitely should discuss this), and then have `cosa
run` just wrap it.

Closes: #371