cmdlib: fallback to virt when on overlayfs
Really, we can support overlayfs fine if we just do the compose in virt.
This is important in CI contexts where we're running in a container. At
least in Kubernetes, one could use an `emptyDir` mount to work around
this.

A more generic approach here would just be to test xattr capabilities
given the target filesystem and environment we're in.

The underlying context for this is that I actually don't want to run a
full compose at all, but rather just be able to update the lockfile
easily using `cosa fetch --update-lockfile --dry-run`. This in fact
doesn't actually require any privs at all, but it'd get blocked by this.