cmd-fetch: add --dry-run
It's useful sometimes to just fetch the latest metadata to see what
packages would be pulled in.

Another use case is in combination with `--update-lockfile`, so that we
can still update the lockfile without having to download the packages.

I debated a bit whether to just add a `-- "$@"` semantic where all the
remaining args are passed to `rpm-ostree`. I think we could do this,
though this specific option is useful enough I think to be a proper
`cosa fetch` one.