Add support for 'rootfs: verity' (and use for /boot if possible)
I'd like to move in the direction of using fs-verity.  First, if
we detect the target kernel has `CONFIG_FS_VERITY`, we use it
unconditionally for `/boot` since that's already ext4, and we don't
need reflinks there.

To aid further development work here, add `rootfs: verity` as an option
in `image.yaml`.  But this isn't the default because we don't
want to trade off incomplete security for performance (reflinks).