create_disk: Use target EFI binaries
This is analogous to
https://github.com/coreos/coreos-assembler/pull/652

We were hardcoding `fedora` here, pulling the EFI binaries
from the coreos-assembler container.  Change to installing
the *target* EFI binaries instead.

First, I verified this produces the same output
and `cosa run --uefi-secure` still works with FCOS.

But, this fixes `cosa run --uefi-secure` with RHCOS, and ensures
we're installing the RHEL8 EFI binaries.

Conceptually now, the only "traces" or "influence" that the
coreos-assembler content has on the target system will (should) be
in the created filesystem structure (i.e. what `mkfs.xfs` does on
disk), but not the actual *contents* of any files.

Closes: https://github.com/openshift/os/issues/383