cmd-build: add --tag switch
I've become quite fond of `cosa tag` as of late to help with keeping
track of which build has what modifications and ensuring they don't get
pruned.

Let's take it one step further by teaching `cosa build` to tag a build
right after it's built. One other tiny advantage of this is that
scrolling back in the terminal, it's easier to see the output of which
`cosa build` invocation corresponds to which modifications.