dasd: allow building and running
This requires libguestfs support for adding custom blocksize for image
file. The patch at upstream is :
https://www.redhat.com/archives/libguestfs/2019-November/msg00025.html

This patch or its later version should go to F30/F31 and eventually be
used by the cosa container.

We never had the issue with libguestfs/guestfish and custom blocksize on
master because we have moved away from anaconda before DASD support
landed.