run: If not specified, randomly pick bios/uefi/uefi-secure
This way we're testing all 3 at least interactively.

We also dedup some of the uefi/uefi-secure code.

What we really want of course is to do this in kola...
but you know that gets right into the kola-vs-cosa issue
that's still lurking in the background still like the odd cousin you
didn't invite to your wedding...