build: Use cache=unsafe when making disk images
We did this before with `virt-install`; if an image build
fails, we always dispose of the disk image.  Using `cache=unsafe`
means that e.g. the `fsync` invocations that ostree makes
inside the build, as well as filesystem-internal synchronous
writes to the journal and metadata aren't passed through to the host filesystem
by default.  This allows the host filesystem to defer writeback,
speeding up image builds.