openstack: build on s390x
If we run `guestfish --remote sh <command>`, guestfish would run /bin/sh
in the real rootfs - which is an ostree rootfs - and /bin/sh would be
not found. Running with `debug sh` would tell guestfish to run in
supermin's initrd space instead.

This is for helping the case where a config drive (ISO) is used to pass
Ignition config file to guests, in the case of terraform's libvirt
provider.

Related:
https://github.com/dmacvicar/terraform-provider-libvirt/pull/666
https://github.com/openshift/installer/pull/2672

Co-Authored-By: Prashanth Sundararaman <psundara@redhat.com>