kola/qemu: Add a swtpm by default
The duplication in qemu code between cosa and kola needs
to end - this is the start of moving some bits into kola.
Eventually we'll have `cosa run` be a wrapper for something
like `kola qemuexec`.

This bit is the equivalent of
https://github.com/coreos/coreos-assembler/pull/878/commits/55c9b651e5ee94fb4de947a943df213ff3c601eb

We want to strongly emphasize use of TPM devices in CoreOS, so let's provision
one by default.

They can be used for disk encryption, secure secret storage, etc.