cmd-prune: Drop --keep-last-days
This option was well-intentioned, but ultimately because it's not the
default pruning behaviour, and because a developer could sometimes do 20
builds in a single day when iterating, it's not super useful.

So let's drop this and simplify a bunch of logic at the same time. I
would also like to put emphasis on using tags for keeping builds around.