Support blacklisting kola test by arch
Following up on #973, for the kola test runs to be successful on
s390x and pcc64le, the TPM test has to be excluded. An arches option would
be useful to list the arches to exclude. Here is an example:

- pattern: rhcos.luks.tpm
  tracker: none
  arches:
   - s390x
   - ppc64le

If no arch is specified, it means blacklist on all arches