buildextend-rojig: New command
For Fedora Silverblue I want to possibly switch to rojig by default.
There are a few reasons for this; among them that on more "pet"
systems package layering is natural and rojig becomes advantageous
then.

This small buildextend just wraps up the rpm-ostree glue for
this into the cosa model.