buildupload: Don't set Content-Encoding for gzip
https://github.com/coreos/coreos-assembler/pull/862#issuecomment-567069130
This regressed OpenShift on OpenStack installs, which had adjusted
to deal with the lack of `Content-Encoding` and a switch to add `.gz`
to images.

Basically, if we're going to say things are explicitly compressed,
let's not also tell people to *transparently* decompress them because
it's easy to end up with a file named `.gz` that was decompressed.