cmd-compress: skip compressing already compressed files
Even if they're compressed by a different compressor.

(How many more derivatives of the word `compress` can we fit in this
commit message?)

This came up with `buildextend-gcp`, which currently hardcodes `tar.gz`,
because that's what the cloud wants when uploading. But then we'd
recompress it again with `xz` at the `compress` stage, which probably
gains us some space, but having a `.tar.gz.xz` is pretty confusing.