Add `offline-update` command
We need to do some upgrade testing.  My specific goal here
is to be able to use `kola`, which is oriented around providing
Ignition - so we need to do the update fully "offline".  I wrote
the quick/obvious script to do this and ran up against the need
for absolutely horrid hacks if grub2-mkconfig is still needed.

But anyways, it works.