cmd-sign: Publish message in fresh process
Run the process that will publish the message from a new process rather
than just a fork so that we don't inherit e.g. the Twisted state.

This is a quick fix until we rework this differently for better sharing
with other tools in the FCOS pipeline that will need to interact with
the fedmsg bus.

For details, see: https://github.com/coreos/fedora-coreos-pipeline/issues/178