buildextend-installer: error out early if missing src config
In FCOS, we're looking to drop the installer ISO in favour of the live
images:

https://github.com/coreos/fedora-coreos-tracker/issues/203
https://github.com/coreos/fedora-coreos-config/pull/261

AFAIK, RHCOS should be able to follow suit once the new
coreos-installer binary makes it there.

Until then though, rather than entirely drop support for building it, at
least check upfront that the `installer/` directory exists so we don't
do a bunch of work for nothing.