create_disk.sh: Query the host page size when fs-verity is enabled
ext4 fs-verity requires the fs block size to be the same as the kernel page
size, and different architectures use different page sizes. So, we have to
query the host page size and use that as the block size when fs-verity is
enabled.

We also have to take the block size into account in the rootfs size estimation,
since larger blocks waste more disk space with internal fragmentation.