aws-replicate: run in parallel
This is a follow-up to #825. We go back to letting ore upload to all
regions in parallel, but make use of ore's streaming output so that we
do write down the regions that did upload successfully.

Note also that `ore aws copy-image` is idempotent, so even if an error
occurs that somehow prevents us from writing down the uploaded AMIs,
we'll eventually converge on the right final state (esp. now that we
retry by default).

This also bumps mantle to include the required change in `ore aws
copy-image` (https://github.com/coreos/mantle/pull/1162). Shortlog:

```
Colin Walters (1):
      kola/qemu: Fix leaking qemu instances

Jonathan Lebon (1):
      ore/aws/copy-image: stream results back to stdout

Stephen Lowrie (2):
      Merge pull request #1163 from cgwalters/qemu-leaks
      Merge pull request #1162 from jlebon/pr/ore-copy-image-json
```

Doing this in the same commit, since it needs to happen atomically.