kola: Add a --cosa-build argument
This starts the ball rolling on having mantle understand cosa,
which is key for long-term dedup of the projects.

The main motivation for this right now is that by decoding the
JSON into a Go struct, we are effectively doing a schema validation.

After this lands I'll make a separate PR to flesh out the struct
entirely, then change the decoder to error on unknown keys.

This also removes some CL cruft around the default disk image.