cosalib/build.py: Make commitmeta.json and COSA information optional.
cosalib._Build was born because the need for commitmeta.json, archiving
COSA itself, and understanding COSA. However, it is the only _Build
derivative the requires that level of understanding. This change adds
kwargs of 'require_cosa' and 'require_commit'.

cmd-koji-upload: instantiate cosalib._Build with constructor kwargs of
with require_{cosa,commit}=True