run: Support --user root
For FCOSB, we don't have a `core` user by default (since we always
expect gnome-initial-setup to make one) in the usual path.  And
right now `cosa run`'s Ignition isn't automatically adding the `core`
user it says to the `wheel` group.

In a lot of cases it's much more convenient to directly log in
as the `root` user anyways.  So let's support `cosa run --user=root`.