Don't enable verity for /boot by default
This breaks the use case of `coreos-installer` being run
from an older system like e.g. Debian stable.

We're not actually making use of verity yet, so let's back
off from that, keeping the code as an option:
`bootfs: ext4verity` in `image.yaml`.

We will revisit this when the verity research plays out more.