schema: Mark container-{config,image}-git keys as optional
I install coreos-assembler inside my existing toolbox so that
I can have one place to develop on e.g. rpm-ostree and other
tools.  This needs to be supported, even if not a primary path.

In this scenario we won't have the bits from `Dockerfile` builds,
so remove those from the required metadata keys.