cosalib/build: Use workdir tmp/ as tempdir
We changed this behaviour during the refactor. There's quite a bit of
history here on why do this. But a major one at least is that we want to
be able to just `rename(2)` the final build artifacts into place. This
saves a bunch of time and I/O.

I noticed this due to the fact that we were losing sparsity from the
output of `qemu-img convert` because `shutil.move` doesn't do the
equivalent of `cp --sparse=auto`. This patch fixes that, though I think
we should also be able to change that call to a simple `os.rename()` in
a follow-up to make it explicit.

Related: https://github.com/coreos/fedora-coreos-tracker/issues/361