Support coreos installer test for non-x86 arches starting with ppc64le
This is the first PR for non x86 arches support for the coreos installer test. This is mainly
needed for the pipeline to build rhcos and so ppc64le and s390x are the main focus. Used grub for
netbooting ppc64le since I am not sure about pxe support in ppc64le although there are some conversations
I have had with folks who seem to think there is pxe support.

TBD: s390x does have pxe support although it's taking a bit of figuring out , but this PR should set the stage for
easily adding s390x support - which I am investigating.