kola: Add testiso command
This currently just tests the legacy installer.  I translated
the existing shell script into Go.  The next steps here
are to clean this up and refactor the code so it can test both
installation paths.

Also replace `cosa test-coreos-installer` to call out to this;
in the future when `kola` learns how to parse cosa build streams
we can delete the old shell wrapper.

Co-authored-by: Prashanth Sundararaman <psundara@redhat.com>