kola: Translate from Ignition spec v3 to v2 as needed
Duplicating Ignition spec 3 and 2 code across our tests is
a huge demotivator for writing tests and is prone to error, etc.

Use https://github.com/coreos/ign-converter
in the case where we're booted to an OS requesting spec 2,
and we only have spec3 in the test.

This is prep for working on install tests, where I didn't
want to write Ignition configs twice.