sdk: Trim away lots of CL stuff
The code in `sdk` closely matches what we do with coreos-assembler
in the concept of a "workdir".  Trim out the CL-specific stuff
like `REPO_ROOT` environment variables, etc.

This is prep for reworking it to detect things like `builds/`
from coreos-assembler.

Now, the "Fedora Cloud Base using plume" code *might* have
used that but I doubt it.  It basically boils down to "pwd" so
let's just assume that in that code path for now.