schema: simplify and drop patterns and "anyOf"
In looking to teach `ore` how to directly interact with the schema, I
found that there is a huge a problem with cross-language support for
"patterns" (the GoLang validators only support a subsect) and then the
"anyOf" and "if/then/else" support is spotty at best.

Further, I came to find that the Python validator doesn't _really_ work
as we expected and is buggy.

In light of needing to support GoLang and Python, simplifying the schema
seems to be the best course for the time being.