kola/spawn: add --reconnect switch
Right now, `kola spawn` immediately kills the node once we leave the SSH
session. This is nice most of the time for quick testing. But it
basically throws out the window anything that requires rebooting (which
oddly makes it more limiting than `cosa run`, which does). And sure,
there's always libvirt for more persistence, but I think there's room
for another gradation level here.

Add a new `--reconnect` switch which can handle this. In this mode, we
constantly try to reconnect to the node after getting disconnected. One
can simply do Ctrl-C to kill the node.