kola/qemuexec: add --inject-ignition
Right now, whether to inject Ignition in `/boot` is auto-detected based
on whether the basearch supports `qemu_fw_cfg`. But we also want to be
able to use `kola` to run metal images as a sanity test.

Add a switch to force this behaviour on in those cases and teach `cosa
run` to use it for metal images.

Related to this, see also https://github.com/coreos/ignition/issues/928.
We'll want to clean up that auto-detection in the future so that instead
kola passes the Ignition config via a separate block device.