Allow having only an overrides lockfile
Right now, we ignore the override lockfile if a base lockfile isn't
present. This mostly works, but it prevents the case of bodhi-updates
where we want to override only some packages but leave the "base"
unlocked.

Tweak things so that both the base and override lockfiles can be used
independently.