mantle/schema: use `schematyper --ptr-for-omit`
Golang doesn't have a distinct `Option` type, so when deserializing
data, it has to create some value of that type. But then when
deserializing, `omitempty` on the struct doesn't kick in because it is a
valid object, even if all its fields are e.g. `""`.

The hack that's used to work around this is to represent optional
structs as pointers. We can tell `schematyper` we want that behaviour
via `--ptr-for-omit`.