mantle/sdk: Clean out version.go
I'm trying to make `mantle/sdk` be a coreos-assembler build metadata + mantle
integration point, since that's a lot of the role that it played
with Container Linux.

Clean out `version.go` in prep for that; only one API call is
still used, and we can clean that up in the callers after.