Teach mantle/kola how to automatically find cosa builds
The current `mantle/cosa` directory is a bit circular, we
basically have `cosa/mantle/cosa` then.  It made sense before
the projects were merged but less so now I think.

Second, the `mantle/cosa/builds.go` is really about the `meta.json`
file but we also need the *directory* for a build.  What's
going on in kola is quite weird because we basically need
to keep around the pair of (directory, parsed meta).

Enhance the `mantle/sdk` to learn about a `LocalBuild` struct
and teach kola how to use it to synthesize the other two if
`.` looks like a coreos-assembler root.

This drains more logic from `cmd-kola` in cosa, trending
towards removing it!