mantle: Rework various bits to use new RpmArch() API
Fedora/RHEL are *defined* to be self hosting; which means
cross compliation is not at all the default.  See also
https://blog.verbum.org/2012/10/13/building-everything-from-source-vs-self-hosting/

The mantle code was just using Portage architecture because
of CL; unifying on the RPM architecture names as are already used
by cosa will help consolidate things.

Further, we don't need to pass a "Board" all the way down
the stack; since we don't support cross compliation there's no
reason not to just call the new global API to get it.