cmdlib: Use kola qemuexec
And this final commit gets rid of almost all the core qemu logic in
cmdlib, so e.g. architecture specific changes only need to happen
in one place.  The only special casing left is `$devtype` and
the terminal.

We inherit the virtio-rand setup, CPU calculations etc.

And now we can just use virtio for the root disk
and the cache disk, which drops all the eyeball-glazing
scsi incantations.

(Well, of course arch changes also end up in libvirt which we
 use indirectly too, but at least *we* only have one place)