Add tang server for encryption tests
Tang is needed for Network Bound Disk Encryption tests.  The
Tang rpm installation uses systemd to create the socket but
systemd in a container requires the host to have seboolean
container_manage_cgroup=true.  Since we cannot guarantee that
is the case on all hosts running coreos-assembler, xinetd
can be used to create the socket.  The one caveat is that tang
needs a wrapper script to redirect stderr so it won't cause
errors to clients requesting keys.