aws: provide way to select source-region
When replicating AMIs across regions, the default is to use the first
region + AMI ID pair in the list of AMIs in `meta.json`.  This is
typically `us-east-1`.

When trying to replicate AMIs to AWS China regions, the RHCOS pipeline
uses a separate set of credentials that does not have the ability to
copy the AMI from `us-east-1`.  In this case, we want to be able to
specify a source region in AWS China where the AMI can be copied from.