mantle/qemu: Support catching initramfs failures
Pairs with https://github.com/coreos/ignition-dracut/pull/146

This way, we error out fast if something went wrong in the initramfs
rather than timing out.  And further, we get the journal as JSON,
so we can do something intelligent in the future to analyze it.

And add a test case for this.