mantle/*/packet: rework for FCOS
Unlike CL, FCOS allows us to override the platform ID of the PXE image.
That allows us to put the installer Ignition config into userdata, which
lets us drop the requirement to upload files to a remotely-accessible
URL.  Instead, allow the user to pass an URL to an iPXE script (which
references the installer kernel/initrd) and an URL to the image to be
installed.

By default, mantle will use an iPXE script URL that points to the cosa
GitHub repository, and an image URL fetched from FCOS stream metadata.
This means we'll use a static installer image, but that seems like a
reasonable tradeoff in order to avoid implementing a webapp or requiring
mantle to upload a file.

Drop the "board" abstraction within Packet code and use Fedora arch
names.  This means we have to call amd64 "x86_64".

Drop various CL-specific workarounds.