grub: drop default `rd.neednet=1 ip=dhcp,dhcp6` kargs
We want to move to a model where networking isn't unconditionally
brought up, but instead only if Ignition requires it.

But I think we'll still have to keep supporting
`ignition_network_kcmdline` since it's kind of part of our API now that
it can be overridden via `ignition.firstboot`.

Works with:
https://github.com/coreos/fedora-coreos-config/pull/321
https://github.com/coreos/ignition-dracut/pull/164
https://github.com/coreos/ignition/pull/956