buildextend-live: Also filter out rhcos.root kernel argument
We don't technically *need* this anymore; I've changed RHCOS
to fully use the `-diskful` targets.

But, filtering out that karg would have also fixed the core
bug of the LUKS bits activating on a live system,
and it's just cleaner to have it filtered out.

(I wrote this before fixing RHCOS to use the `-diskful` targets)