mantle/kola/tests: Add Tang and SSS encryption tests
Add tests for Tang and SSS to test encryption on RHCOS.  The Tang
test requires Tang to be installed on the host running the test.
It will gather the thumbprint directly from Tang and use it in the
ignition config.

xinetd is used to manage the socket connection to Tang.  A modified
xinetd.conf file is used to run unprivileged.  xinetd is run by
the user running the test and re-uses the webcache ports defined
in /etc/services.