mantle: Remove generated cosa schema from git
My recent change to use `make` for this seems to have introduced
a race condition mainly in Prow; perhaps something is not
preserving timestamps.

In the end anyways, files stored in git should not be touched
via a default `make`, as that is both subject to timestamp
problems and also creates friction.

Generating it is cheap, so let's switch to always generating it.