build-sys: Require explicit `make schema-update` to regenerate
Alternative to https://github.com/coreos/coreos-assembler/pull/1306

My recent change to use make for this seems to have introduced
a race condition mainly in Prow; perhaps something is not
preserving timestamps.

Files stored in git should not be touched
via a default make, as that is both subject to timestamp
problems and also creates friction.

So this change now requires an explicit manual generation; and
we also consistently add the generated doc file to git too.