mantle/ore: gcloud: remove bucket subpath from GCE image name
Before this commit if I try to upload to gs://mybucket/fedora-coreos
my image name would get rewritten to append the path `fedora-coreos`
into the image name. For example:

- `fedora-coreos-31-20200404-dev-1-gcp-x86-64`

turns into

- `fedora-coreos-fedora-coreos-31-20200404-dev-1-gcp-x86-64`

This was introduced in d98ec2e and doesn't seem to be needed.
Let's rework it so that transformation doesn't happen.