kola/cluster: upload files to temporary path and rename
As good practice, upload to a separate file first before renaming it to
its final location. That way there's no chance of the file sticking
around in its official spot even if we fail. But also, other services
that are watching the path won't get a partially written file.

Prep for next patch which also adds labeling to that.