fcos.upgrade.basic: sync after untarring OSTree tarball
This works around a subtle issue in rpm-ostree/libostree dropping D-Bus
transactions due to `mount` causing a cache flush and hanging for a
while due to slow I/O.

As mentioned in the comment there, we should drop this in the future and
rework things in the stack proper instead so we're not susceptible to
this.

See https://github.com/coreos/coreos-assembler/issues/1301.