build: Identify target disk by wwn
This is a followup to:
https://github.com/coreos/coreos-assembler/pull/1342
which is part of:
https://github.com/coreos/coreos-assembler/pull/1289

Our builds right now run qemu in a way that's highly sensitive
to changes in device ordering.

SCSI hardware supports a [world wide name](https://en.wikipedia.org/wiki/World_Wide_Name);
change our qemu invocation to use `42` (in decimal) for that.

Unfortunately because we're not using udev in our supermin
VM right now we don't get the nice `/dev/disk/by-id` symlink
to it.  Instead just walk `/sys/block` manually.