cli: always call startLogging in WrapPreRun
There is an open issue in spf13/cobra where if a child possesses a
PreRun or PreRunE function the parent's functions are ignored. As we are
only starting logging at the top level command this causes the logging
parameters to be ignored in multiple commands today. Instead just always
call startLogging inside of the WrapPreRun function.