cmd-build: add --strict for building in strict mode
This makes use of rpm-ostree's new `--ex-lockfile-strict`. We really
want this for FCOS CI and release builds. For more background, see:

https://github.com/coreos/fedora-coreos-tracker/issues/454

This also drops the printing of lockfiles and the `sleep 1` bit,
since... I think by now people should be pretty familiar with lockfiles
and rpm-ostree prints them already (and it's easily viewable while
libsolv is loading its cache).