cmd-build-fast: New command to iterate on content quickly
Part of the implementation of
https://github.com/coreos/coreos-assembler/issues/1341

This is useful for me hacking on ostree/rpm-ostree.  Here we
don't need to regenerate the initramfs (usually), we just want
new ostree binaries in `/usr`.  But we *do* want the OS to be
set up as a "proper" OSTree commit, i.e. not `rpm-ostree usroverlay`.

Here this completes a qemu image (overlaid on previous) in 11 seconds,
compared to 147 seconds for a full build; ~13x faster *and* we're
not writing `O(sizeof(coreos))` bytes to my SSD, but
`O(sizeof(change) + sizeof(metadata(coreos))`.

With this a workflow looks like:
```
$ cd /src/ostree
$ make && make install DESTDIR=/path/to/cosa/overrides/rootfs
$ cd /path/to/cosa
$ cosa build-fast
$ cosa run --qemu-image tmp/fastbuild/fastbuild-qemu.qcow2
```

(Or, instead of `cosa run`, `cosa kola run --qemu-image fastbuild/fastbuild-qemu.qcow2 -E /src/ostree ext.ostree.*`)

You can also use this inside a project directory by setting `COSA_DIR`.