platform/qemu: Enable forwarding multiple ports
One of the limitations of unprivileged qemu is there is no
machine to machine networking.  Host port forwarding allows the
host machine to communicate to the guest machine through the local
host address. It also allows the guest machine to communicate
with the host machine through the gateway address. With two
guest machines, we can bridge the two guests through the host
machine with port forwarding. For example, Host A can run a
webserver on port 8080 and Host B can access Host A's webserver
through its gateway (host machine) address http://10.0.2.2:8080.

This is useful for creating test services (such as tang) for use
by the test machine.