build: Don't auto-load qemu image when building
Regression from https://github.com/coreos/coreos-assembler/pull/1289
that probably explains other mysterious problems I was seeing too.

Basically, if a qemu image already existed, we would auto-load it
with `kola qemuexec`.  In the case of builds, we absolutely do not
want to do that.

I probably didn't see this *initially* when working on this because
my build didn't have a qemu image yet.