cmdlib: Don't dump serial console to stdout
In https://github.com/coreos/coreos-assembler/pull/1338
we fixed the problem of having one's terminal be corrupted
by the serial console.

This fixes the same problem for `cosa build` which ends up
running a VM too.  First, there's no real reason to dump
the kernel bootup output to the terminal (or build logs) - it's
totally uninteresting stuff that we only want to see if
something fails.

Gather the serial console to one log file, and the output of
the *command* we run in the VM to another.  Fork off a
`tail` process to follow just the command output.  (When
we rewrite this in a better language we will do this properly
async)