qemu: Unlink non-multipath/nbd disks
I ran out of space in `/tmp` today and it turned out I had
a whole lot of `/tmp/mantle-qemu` files, including some
really big disk images.  This is a combination of changes
from https://github.com/coreos/coreos-assembler/pull/1338
and the multipath PR https://github.com/coreos/coreos-assembler/pull/1296

Basically if we get interrupted by Ctrl-C which is a lot more
likely now when qemu isn't owning the serial console from the
start, we will leak our tmpfiles.

We probably need to install a `SIGINT` handler but that's a whole
mess that's avoided if we just let the kernel clean up our
resources for us, which is what was happening before the multipath
PR (well, at least for files and not the swtpm tmpdir, but that's
small).

And actually
https://github.com/coreos/coreos-assembler/pull/1380
does install a `SIGINT` handler so someone please review that
and we can maybe fix this more after that lands.