cmdlib: let qemu write command output on its stdout
Instead of using `tail -qF`, just write the output to a virtio-serial
port and tell qemu to output that on its stdout. That way, we don't have
to worry about cleaning up the `tail` process at all.

This also fixes an issues we've been seeing where `tail` is killed
before it had a chance to print the entire command output, which
crucially might omit the final error message.

Note we close stdin to qemu here otherwise it seems like it just waits
forever for input to be written into the virtio port.