cmd-fetch: add --with-cosa-overrides
By default, `cosa fetch` ignores overrides in `overrides/rpm` since
those are meant to be temporary and might be e.g. locally built
packages. (Otherwise, for example, removing a cosa override might cause
`cosa build` to fail since the no longer overridden package was never
actually downloaded).

However, it can be useful sometimes to not ignore it if we know that
it's not a "temporary" package and that it won't be found otherwise from
the regular repos. This is the case in the CI of `fedora-coreos-config`
where we use `overrides/rpm` to test composes with overrides without
having to tag them into the pool first.

Related: https://github.com/coreos/fedora-coreos-config/pull/365

Closes: #1117