kola: rename --tags to --tag and tweak help
The name `--tags` makes one think that it takes a list of e.g.
comma-separated tags. Rename the switch `--tag` instead while it's still
fresh and specify in the help string that it can be given multiple
times. This is the same as e..g `--blacklist-test` and
`--debug-systemd-unit`. Tweak their help strings as well.