kola/testiso: test that default path is fully offline
I'd like to get coverage for
https://github.com/coreos/coreos-installer/pull/197 before it merges and
we do a release with it.

To do that, I want to test that the default installation path no longer
requires Internet access. Here, "default" means we pass neither
`--offline` nor `--stream` nor `--image-url` to
`coreos-installer install`.

Change the new `iso-offline-install` scenario to cover this; all we need
to do is remove the explicit `--offline`.

This scenario will only pass for now if the build under test has that PR
and the live ISO has the osmet files (which right now is gated behind
`--osmet`).

Once that PR is in FCOS proper, we can make `--osmet` the default and
this test will then always pass.