installer: Refactor initramfs code
This is a code cleanup patch in prep for further work
to move the rootfs squashfs into the ISO, which
is part of what `fulliso` was doing.

With this patch, first we extract a `mkinitrd` function
which takes a whole directory contents, which avoids
the fragile "pass file list to cpio" that was duplicated
in a few places.

It's also a lot cleaner if we structure things so that
we first generate the "base initramfs" + "stamp initramfs"
which is common to both live and legacy, and then in
the live case further append osmet/rootfs.