installer: Put rootfs on ISO
For FSBCOS¹ a while ago I added the `fulliso` variant here
that was removed with the osmet rework.

`fulliso` did two things:

 - Embedded the "metal" image inside the ISO
 - Moved the rootfs squashfs out of the initramfs and into the ISO

The osmet work obsoleted the *first* bit in a much better way,
but the second problem still remains - the initramfs generated
for FSBCOS is just too large to boot.

But can have the best of both worlds; for the ISO case,
always put the rootfs on the ISO and not the initramfs, because
we still have the code to mount it in fedora-coreos-config.  It
makes us start the initramfs noticably faster too.

For the PXE case, we continue to embed the squashfs in the initramfs.

But for Silverblue I don't care about PXE (or if you want to do that,
PXE boot FCOS and use `coreos-installer` to install FSBCOS!).
So add a `--no-pxe` option for this so we don't pay the
cost of generating a huge initramfs that won't even work anyways.

¹ Fedora Silverblue based on Fedora CoreOS; https://github.com/cgwalters/fedora-silverblue-config