fcos.upgrade.basic: only force updates on if required
On production streams, we should never have to force Zincati updates to
be on since that should already be the default. So remove that from the
Ignition config and dynamically check this when testing.

We check twice here because we need to account for the case where we're
first rebasing from e.g. `testing-devel` to `next` (where we do have to
force updates) and then an update on `next` itself (where we shouldn't
have to force updates).

This would have caught a regression in `next` where Zincati updates were
disabled: https://github.com/coreos/fedora-coreos-tracker/issues/473.