Add cosa supermin-shell
While working on the oscontainer stuff in supermin I really
found myself needing a shell in the supermin environment.
Add a handy command for that.

In order to share code between the two paths, rework
the supermin argument handling to build up a single
array of extra arguments for qemu.

(Hopefully it won't be long before we replace this with
 something saner like running a proto-CoreOS w/systemd
 and using virtio-fs for the rootfs)