cmdlib: Lower cost of cosa RPM overrides repo
Surprisingly, AFAICT there is nothing in libdnf today which
automatically lowers the cost of local repos vs remote ones. Let's give
it some help by adding a `cost=500` here (the default is 1000). This
will allow libsolv to always pick the RPM from the overrides repo, even
if the same NEVRA is available from a remote repo.