buildinitramfs-fast: New command to iterate on initramfs quickly
This is parallel to [build-fast](https://github.com/coreos/coreos-assembler/pull/1371)
which is about quickly iterating on userspace.

Here is the opposite - we just want to take our new code and
get it into the initramfs as fast as possible.  An example use
is demonstrated in the comment at the top of the file.

Here this takes about 5 seconds to run, which is fast enough
that at least I would probably stay at the terminal and not e.g.
context switch to code review or check Twitter or whatever.