kola: Add option to configure RAM
Currently we default to 1024MiB for tests, but trying out this
patch with:
```
$ cosa kola run -- --parallel 8 --qemu-memory 512
```

The *only* test which fails is `rpmostree.install-uninstall`, probably
because the Fedora package metadata is enormous.

The main thing limiting test parallelism in our CI pipeline is RAM,
so if we can e.g. use 512 by default then that would nearly double test execution
speed.

This is prep for a future patch which will try to assign "instance type requests"
to tests, so both qemu and other platforms like AWS can use smaller instances.
For example in AWS we could use `t2.nano` for some tests instead of `m4.large`;
the pricing difference is large.  `t2.nano` in `us-east-2` is $0.0058, and
`m4.large` is $0.10, a savings of 95%.