mantle/ore: gcloud: add promote-image command
In our build/release workflow for GCP we build and upload/create
a GCP image in the build pipeline. We deprecate the image (because
we haven't decided to release it yet) and then add it to the
corresponding image family for the stream it was built against.

The promote-image command will be used in the release pipeline to
undeprecate the image now that we have decided to release it. Since
we only want one image visible in an image family at any given time,
the promote-image command will then deprecate all other images in
the specified image family.