build: Add support for "variants"
My immediate goal here is to change
https://github.com/cgwalters/centos-coreos-stream-config
into something more like "openshift/coreos-config" that
can be built in multiple ways.

We'd have `rhcos` and `cscos` as "build variants".  And
actually while we're here potentially `rhbasecos` which would
be "just stuff from RHEL".

Another use case for this is the "alternative desktops" for
Silverblue.

And yet another use case is something like a "debug" build
for FCOS where e.g. we swap in `kernel-debug` and potentially
turn on other userspace debugging too.