oscontainer: use workdir as temporary directory if provided
So that e.g. buildah uses it for temporary storage and doesn't blow
through the tiny supermin appliance rootfs (and sure... we could just
bump that from 5G, though that bumps it for everything else, and I think
it's cleaner to take `--workdir` at face value).

This should resolve issues the RHCOS pipeline is hitting where `buildah
commit` is hitting `ENOSPC`.