cmd-oscontainer: put workdir in cache qcow2 in unprivileged path
We were passing `--workdir /host/container-work` to `oscontainer.py` in
the unprivileged case, but that path isn't actually mounted from the
host.

And actually, even if it were mounted, buildah wants to be able to
`chown` things in its container storage, which it won't be able to over
9p.

Instead, let's just point it at the cache qcow2 to do its work. That
way it can do whatever privileged operations it wants there, and plus we
get to cache the base `--from` image this way. Other temporary files are
nuked on exit, and otherwise, on the next run of `oscontainer.py` (in
practice, pipelines don't currently cache the qcow2 in between runs
anyway, but we should be nice to the unprivileged developer path).

(Again, we could also bump the supermin root instead, but I think as a
general pattern we should leave it tiny and do the real work in a
mounted disk which has lots of space and a more modern filesystem than
ext2.)