kola: Add RHCOS LUKS upgrade test
I really think osmet is key for the RHCOS Live ISO experience and
we don't have that right now because of the LUKS container
thing.

I thought then about putting the LUKS header in `/boot`
and dropping the LUKS container by default
(that came up before) so we could also skip the dm-linear
thing.  That way it'd be the same as FCOS.

I started looking at that code but then I was reminded of a
moment of fear I had when something related came up last week -
we have absolutely no testing of upgrades around it, and that
is *highly* likely to break if we start trying to make any
fundamental changes.

This paves some of the infrastructure for a bit more "real" upgrade
testing where we boot the old OS with a target Ignition (in
this case the LUKS config), then reboot into the new OSTree
commit.  This is in contrast to the "synthetic" `offline-upgrade`
which directly overlays the new OSTree commit.

Run this test via e.g.:
`kola run-upgrade --ignition-version v2 -b rhcos -v --find-parent-image --qemu-image-dir tmp/`
which needs to be plumbed into the pipeline.