buildextend-live: use `find . -mindepth 1` for cpio
Super subtle but evil regression from #1423. By switching to `find` to
feed filenames to `cpio`, we were passing it `.`. And `cpio` dutifully
made note of the permissions of `.`, which at the time it is run are
those of the temporary directory we allocated, which is naturally 0700.

This in turn breaks Ignition's `files` stage, which purposely runs
things like adding SSH keys as the target user, and thus it can't even
access the directories in `/` when doing the equivalent of `mkdir -p
/sysroot/var/home/core/.ssh/...`.

Instead, use `-mindepth 1` here so we always skip the root of the
temporary directory itself.

This resolves https://github.com/coreos/fedora-coreos-tracker/issues/496
but not marking as `Closes:` because we should really add a basic live
PXE + SSH key test.