ci: run on Fedora 32 image
Now that we've bumped cosa to f32, we want to make sure CI is also
targeting that. (Ideally, we'd just build from the Dockerfile and run
tests on the resulting image. That's something we can easily do with
Prow, though we should be able to do this via the OpenShift Jenkins
plugin too.)