oscontainer: Use podman push --digestfile
A while ago, we were using `skopeo inspect` as part of `pivot`
for OpenShift 4 OS upgrades, and it turns out that enumerates
all tags, which is expensive and rate limited by quay.io:
https://github.com/openshift/pivot/pull/51#

The RHCOS oscontainer builds have been failing in `skopeo inspect`
and I suspect it's related to this.

It looks like nowadays podman has `--digestfile` which is
exactly what we want here, so let's use it and drop the `skopeo`.
This simplifies the code too.