testiso: make iso-offline-install a default scenario
With osmet in place everywhere, it's expected that it will be the
primary installation method when booting the ISO. To reflect that, let's
make `iso-offline-install` a default scenario, and make `iso-install` an
opt-in instead. So now the default scenarios cover both major "axes" of
installation (PXE/online and ISO/offline).