cmd-build: add --parent-build to replace --parent
Right now, the FCOS download page is using the pkgdiff from the
`meta.json` files to print the changelog. However, this diff is against
the previous build, not the previous release. This distinction is
important since e.g. we might do multiple `testing` builds, but only
release one.

Ideally, the JS client code would perform the diff, though parsing and
comparing RPM versions and getting this correct is no fun.

Let's add a new `--parent-build` which replaces `--parent`. From this
build, we extract the OSTree commit to use as parent, as well as the
commit object itself for performing the diff at the end.