metal: don't redirect coreos-installer output to console
In trying to debug why the pipeline metal4k offline install tests were
failing, neither the console nor the journal logs were very useful. The
first because ISO installs don't output to serial by default (and we
currently don't have an easy way to override that), the second because
we were redirecting `coreos-installer` to the console, which we can't
capture.

Drop the `StandardOutput`/`StandardError` overrides so that output goes
to the journal, where we can collect it over virtio.