WIP: Enable parallel buildextend-* execution
I really want to speed up the FCOS and RHCOS pipelines by parallelizing
things more. For example, all the simple "guestfish && tweak platform
id" buildextends could easily be parallelizable.

Really, the critical bits that need locking is updating `meta.json`.
Everything else can be done in parallel.

This starts the ball rolling for `buildextend-metal`. We adapt the bits
from `cmdlib.sh` and `create_disk.sh` which currently assume exclusive
access to some directories so that they can run in parallel, then add
locking in `buildextend-metal` at the crucial part where it needs to
write to the build directory.

With this, I can successfully run `buildextend-metal` and
`buildextend-metal4k` and `buildextend-qemu` in parallel.

If agreed on the model, I'll adapt the remaining commands for parallel
execution.