mantle: use better terminology for exclusion/inclusion
The use of the term "blacklist" to describe things we want to exclude
is archaic and not representative of the social values we want to
support in this project.  Similarly, "whitelist" as a term for
inclusion has the same problems.

This changes the code to use a format of `denylist` and `allowlist`
across the project.  It does not attempt to change any of the
vendored modules.