kola: switch qemuoptions to bios for all arches except aarch64 and x86 metal4k
The pipeline for ppc64le fails on the metal4k tests as UEFI is not supported.
Change to use UEFI only for aarch64 and x86 metal 4k test.For s390x, this test is
skipped but in the pipeline it fails because there are no scenarios to test.
Change that message to just a print.