koji-upload: Use builds/ for tempdir
Koji/Brew forces us to recompress files which uses a lot of
space and is exacerbating disk space issues with the RHCOS pipeline.

We want a broader fix where we generally write directly into `builds/`
for images but for now, this quick hack should get us using
a larger PVC for the RHCOS pipeline.