Fix vsmt value in libguestfish for Power
The libguestfish calls need to have the vsmt fixed.

A brief history why we need to fix it:
In QEMU (4.2.0) the vsmt was set to smp_threads by default (29cb4187497d).
RHEL7 doesn't allow vmst to be set due some missing backports, this change
is only possible starting from RHEL-ALT/RHEL8. Our pods are running POWER8
with RHEL7. To fix it we need to restore the vsmt value for default in POWER8.

On POWER8 the vsmt default is 8, when on POWER9 it is 1.
It is also compatible with POWER9, since POWER9 also supports smt 8.