kola: Extend reboot support by implementing "prepare" API
I want to add testing of forced poweroff (i.e. `reboot -ff`)
into OSTree's test suite so we can reliably verify that
it lives up to what it claims on the tin.

The autopkgtest already has a "prepare" API which is exactly
what we want here.  Implementing this though is somewhat
tricky because we need to synchronously get the mark (i.e. state)
out to the harness.  I added a big comment about this.