kola/rhcos/luks: pull tang container from quay.io
There have been intermittent failures in the Tang/SSS tests starting
the Tang container.  The error message is a bit cryptic saying the
Process exited with status 125.  This commit will print out stderr
to facilitate debugging.  The error is suspected to be caused by an
issue pulling from docker.io.  Lets switch to quay.io to see if that
alieviates the failures, if not stderr should give more clues on
the failures.