kola: add minMemory in kola.json (external tests)
For the root reprovisioning work, we need to be able to temporarily save
the rootfs into RAM. This means we need at least 4G of RAM.

Add a new `minMemory` field which tells the platform to make sure the
machine provisioned has at least that amount of memory. On QEMU, this
directly controls the memory of the machine. On cloud platforms, it
would affect the instance type to use. But for now, we only actually
handle this field on QEMU.