cosalib: lock meta.json before reading and writing
This imposes a 15s lock on the cmdlib.load_json and cmdlib.write_json
functions via an NFS safe lock. In order for the lock across namespaces
(e.g. different pods) and read by other COSA processes, the lock is
written to the same path as the target file.

Implements: cosa-20200804
https://github.com/coreos/enhancements/pull/2
Signed-off-by: Ben Howard <ben.howard@redhat.com>