cosalib: lock meta.json before reading and writing
This imposes a lock on the cmdlib.load_json and cmdlib.write_json
functions via an NFS safe lock. In order for the lock across namespaces
(e.g. different pods) and read by other COSA processes, the lock is
written to the same path as the target file.

The lock will lock until released with a max lifetime of 15s.

This is prep for adding callbacks to `write_json` for doing more
complex operations under the lock such as merging.

Partially Implements: cosa-20200805
https://github.com/coreos/enhancements/pull/2