cosalib/meta: support for meta.json merging
Instead of blindly overwriting meta.json, COSA now supports writing to:
- meta.?.json: where ? is the artifact name, or a timestamp
- merging with the on-disk diversion

Merging is done via `write_json` with a merge_func callback.
This enables the lock to be held in one spot. `write_json` calls the
`merge_func` with the on-disk version with the version being merged,
preserving the lock. The version with the higher stamp is the base
version for the merge.

Partially Implements: cosa-20200805
https://github.com/coreos/enhancements#2