create_disk: Use bootupd to install UEFI if available
See https://github.com/coreos/bootupd
and https://github.com/coreos/fedora-coreos-tracker/issues/510

Basically in order to handle *updates*, bootupd also
takes care of installation so that it knows the original
version.

In order to sanely "ratchet" this change, only use bootupd if
we find the ostree deployment is using it.