mantle: Avoid two empty string `""` in most Ignition calls
Seeing the two `""` in most of our calls like
`conf.Ignition("").Render("", kola.IsIgnitionV2())`
felt confusing - I had to go look up what those strings meant.

For the `Render()` case the main reason we're still carrying
some CL code is because some of the tests we'd eventually
like to port still refer to it.  But we're not actually
using it, so split off a separate API for the unusual case.

And make an `EmptyIgnition()` API instead of providing the empty
string.