testiso: Support pxe-install test on aarch64
This adds support for running pxe-install test on aarch64. In the process of doing this , I discovered
that boot once is not supported and the boot order had to be changed after installation (unlike s390x where
the disk is always checked first). This got me back to using QMP again and it works pretty well and is pretty
convenient to change the boot order of a running system

One follow up would be to try to use QMP in general for injecting bootindex for pxe and iso test cases and do
away with adding the primary disk for testiso altogether.