Only add cache disk for rpm-ostree builds
A race that's started happening now that we're trying to do
builds in parallel; the cache disk is only relevant today
for the rpm-ostree-in-VM operations, so only provide it for
that run.  This way we won't have multiple VMs mounting
the same disk writable which qemu detects and errors out on.