qemu: Add comment discussing libvirt and links to related projects
Let's answer the "why not libvirt" first.

I think syzkaller is a cool project in particular, and was
curious about how they do kernel testing, so I recently spent
a bit of time looking at their code.

And kata also has a lot of strongly related bits.

So hopefully in the future we can share ideas if not actual code.