qemu: Make primary disk required, don't TFTP etc.
A few times I've ended up with a non-bootable qcow2, and
the behavior of `cosa run` is that you end up hanging at
`[Link:up, TX:0 TXE:0 RX:0 RXE:0]`
which is the TFTP fallback.

Add an API and use it so that for the cases where we know
we're booting a qcow2, we tell the firmware only to boot
from a hard drive, so we end up hanging at:
`Boot failed: not a bootable disk`
which is at least a much more obvious failure.

(I think with QMP we might be able to detect this and
 error, needs investigation)