cmd-virt-install: add --console
With `--console`, we immediately hook into the VM's console once it's
started. This is different from dropping `--noautoconsole` which would
make it want to connect to the VGA (one can use `--nographics` on top of
that to force it to connect to the serial port instead, but sometimes
you do also want to be able to see the VGA output).