exttests: New container image with upstream tests
This is aiming to finally, finally close the loop around
running upstream tests as part of our CI.

We landed support for "exttests" in here for a while and
since then we are successfully running upstream test suites
using kola on upstream pull requests.

But we also want the inverse: run upstream tests on builds outside of PRs to those repos.
For example,
I really want to run ostree's "transactionality" test on FCOS builds so that if a kernel change breaks it, we know that.

In practice, it's likely that down the line we
will need to version at least some of our tests along
with the OS content - by building them as RPMs or
sticking them in containers that are versioned with the OS
or something.

But for now, building fixed upstream releases will
give us a lot of reproducibility/control, at the
cost of needing to periodically bump them.