mantle: fix podman.workflow stop test
Sometime between podman 1.x and 2.x podman started putting
full 64 character IDs into the json output. Dynamically detect
the length of the ID and compare that number of characters.

We haven't noticed this test failing because we've been denylisting
the test in the pipeline: https://github.com/coreos/coreos-assembler/pull/1478