osmet-pack: run coreos-installer in target rootfs
Otherwise, we might get linking issues if the target system has newer
e.g. glibc. We could use `bwrap` here instead of `chroot`, but meh this
is easier and the block device itself is read-only anyway.

I was initially trying to get this code to use the qemu image directly,
but we have a bunch of issues there with our systemd services getting
confused by the metal disk containing identical boot and root
partitions (which... we should circle back to eventually).

Closes: #1761