kola: don't resolve symlinks for ext test naming
Otherwise using patterns in `kola-denylist.yaml` won't always work since
`src/config` might be a repo or might be a symlink.

Let's just standardize on `config` being the name of "the repo which is
at src/config" so that we don't have to think about this when adding
tests to the denylist.