entrypoint: avoid triggering `build` at `install` time
I do `make && sudo make install` and golang wants to rebuild the
binaries as root, which fails because the Go environment isn't set up
there.

There are ways to deal around this (using eg. Colin's `makesudoinstall`:
https://github.com/cgwalters/homegit/blob/735331bb748c71c5b2fac36a10c891381e38e0ae/dotfiles/bashrc#L21),
though it's good to clean up the `install` target to have it actually
key off of the binary. This matches the mantle Makefile.