mantle: work around for ssh crypto policies change
In 5c036d1 we swtched to ecdsa keys for the new crypto
policy change. In fdcfbbc we switched it back because
when we create new ssh keys on AWS we'd get an error because
that platform only accepts rsa keys:
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-key-pairs.html#how-to-generate-your-own-key-and-import-it-to-aws

For now add both a rsa key and a ecdsa key.
Using an ecdsa key allows us to workaround the updated
crypto policies in F33+ that disallow the RSA SHA-1 keyexchange
algorithm, but we still need an ssh-rsa key because AWS only
supports ssh-rsa keys and a key gets generated and added to
AWS every time we start a kola run. We'll add the rsa key
first since platform/machine/aws/flight.go only adds the first
key (keys[0]). We should be able to go back to RSA keys only
when the golang library is updated. More info in:
https://github.com/coreos/coreos-assembler/issues/1772