[4.4]create_disk: remove immutable bit for s390x
The 4.4 pipeline for s390x was failing kola tests because of an updated ostree in the cosa container. This contained the fix for https://bugzilla.redhat.com/show_bug.cgi?id=1867601
which made the root file systems immutable. But the ostree version of RHCOS does not contain the fix causing rollbacks and upgrades to fail.

Removed the immutability bit to fix it as suggested by @cgwalters.

Co-Authored-By: Colin Walters <walters@verbum.org>
(cherry picked from commit 8964218106488f3d72be45cbe8be0fd4c4502e83)