[4.5]create_disk: remove immutable bit for s390x
The 4.5 pipeline for s390x was failing kola tests because of an updated ostree in the cosa container. This contained the fix for https://bugzilla.redhat.com/show_bug.cgi?id=1867601
which made the root file systems immutable. But the ostree version of RHCOS does not contain the fix causing rollbacks and upgrades to fail.

Removed the immutability bit to fix it as suggested by @cgwalters.Also opened a BZ to include fix in ostree for RHCOS 4.5

Co-Authored-By: Colin Walters <walters@verbum.org>