cmdlib: stop copying resolv.conf
This is breaking `supermin --build` for me because it doesn't know what
to do with `etc/resolv.conf` in the prepare dir.

But anyway in user-mode networking, Slirp already takes care of DNS
forwarding in the supermin VM to the host's DNS provider so we shouldn't
need to copy `resolv.conf` at all.