create_disk: move BIOS-BOOT to partition 1 and /boot to 3
The offset of the BIOS boot partition gets encoded into the boot sector,
and that offset is currently 512 MiB, which is awkward if we're going
to support rearranging boot-related partitions.  Move the BIOS boot
partition to partition 1 at the beginning of the disk, and move /boot to
partition 3 on all platforms (including ppc64le, where it was overridden
to partition 2).