Entrypoint: Support for multiple pods and multiple stages
In a previous PR, an unrelated question triggered a line of thinking
that illustrated that `entrypoint` was not being creative enough. This
is the result.

Previously to this change, COSA expected the orchestrator to know enough
about COSA to both setup and run COSA. When running COSA in various
environments this a high bar. This change moves the knowledge of "how
to run COSA" into COSA itself, namely:
- Extends the JobSpec to support "stages". Each stage can run arbitrary
  commands such as "cosa build." Stages are run sequentially, but the
  commands can be run concurrently.
- Full live-cycling of worker pods. Worker pods are created, monitored
  and deleted.
- Defines, but is not implemented, the ability to create non-blocking
  pods for stages.
- Enables serving of files from the buildconfig pod to the worker pods.
  Worker pods can fetch sources and configs from the buildconfig pod.
- Full streaming of logs to the buildconfig and saving of logs.
- Worker pods return the results to starting log via Minio.
- Stages can require an artifact, which requires something (i.e. a
  worker) to place it on the origin pod.
- Stages support prep and post commands.

Worker pods are tightly coupled to the buildconfig pod, by design. A
worker pod is created from the same pod definition of the buildconfig.
`entry builder` determines whether its running in buildconfig or worker
mode by checking for an OpenShift buildconfig spec or a COSA entrypoint
`workSpec`. When a worker comes up:
- initializes /srv
- auto-discovers secrets
- runs the stage(s) assigned to it.
- returns the results over Minio

The JobSpec now supports short-hand artifact creation. For example:
    job:
      build_name: me
      strict: true
    recipe:
      git_ref: testing-devel
      git_url: https://github.com/coreos/fedora-coreos-config
    stages:
     - id: base
       description: "building"
       build_artifacts:
       - base
       - metal
       - metal4k
     - id: live-iso
       own_pod: true
       description: "live ISO"
       require_artifacts:
       - base
       - metal
       - metal4k
       build_artifacts:
       - installer