Entrypiont: add "ci" command
Requiring COSA to run as a buildConfig is not always feasible and it
runs its pod as a fully capable build pod.

For example:
        entry --jobspec jobspec.yaml ci

The pod running entry, must be launched with a service account that can
create pods (i.e. the Jenkins SA). OpenShift deployed Jenkins instances
are calable.

The results of run will be put in /srv. It's envisioned that Jenkins
Pipelines and Prow could use this interface, while developers and ART
pipelines can use the BuildConfig interface. One major difference with
this interface is that cannot use the binary source interface yet.

To ensure parity, the CI mode mock's the OpenShift Build v1 API.

Signed-off-by: Ben Howard <ben.howard@redhat.com>