More draining of create_disk args into JSON passing
Followup to https://github.com/coreos/coreos-assembler/pull/1787/commits/b769f56560b099396f84a9ce694ae3586ddcfe41
All of the parsing and re-parsing of the image YAML into an argument
list for `create_disk.sh` was messy.  Particularly things like
having the defaults hardcoded in the middle of code and using `NONE`
for the empty value, etc.

Add almost everything into JSON that we pass in so we don't need
to "stringify" everything.

The main thing I didn't touch here was kargs but that should also
be better with this because we can cleanly pass an array.