kola: enable tang based LUKS test for all arches
with ignition based LUKS landing, there is no reason why LUKS shouldn't work on other
arches with tang. Just had to build the tang container image similar to x86. Also had to
change the lookup of devices to not use by-partlabel for s390x because of the endianness
issue in naming.