kola: support basic (non LUKS) rhcos upgrade test for all arches
Added a test for the migration of rootfs from crypt_rootfs with dm-linear mapping to the non LUKS root filesystem which will
be the default going forward. This test will only be useful so long as we upgrade from pre 4.7 systems to 4.7 and beyond. This
is good to have for also testing multi-arch upgrades which can now be enabled as part of the pipeline.