mantle: Check for SELinux AVCs just like systemd failed units
We shouldn't ship an OS that boots up and has a SELinux denial
by default; it makes us look less than fully competent.

The motivation for this is
https://bugzilla.redhat.com/show_bug.cgi?id=1900898

Though it turns out we do have a SELinux denial in FCOS by
default now, that's
https://bugzilla.redhat.com/show_bug.cgi?id=1903335
So this can only merge once that's fixed.