create_disk: fix ref handling when not targeting latest build
If we're doing e.g. `buildextend-qemu` against a build which isn't the
latest one, `create_disk` would still deploy the latest one (or more
accurately, whatever the ref is currently pointing to in the `tmp/repo`
OSTree repo). This used to work, so we probably regressed at some point.

Let's just always pull by commit, and then in the remote case create a
remote ref for it.