qemu: automatically inject serial console karg into ISO
This is mostly a revert of
https://github.com/coreos/coreos-assembler/pull/1652, though with a spin
on it: we automatically detect if the `coreos-installer` in the cosa
container has https://github.com/coreos/coreos-installer/pull/341.

This will allow us to test the functionality in that PR's CI before
merging, as well as avoiding ugly warnings until it gets into cosa
proper. (This addresses the original reason for the revert.)

Another difference is that we capture the stderr and make it part of the
warning to make it less scary if operating on an older CoreOS ISO and to
make it clear that it's non-fatal.