buildextend-live: hybridize live ISO for UEFI boot
For BIOS systems, we're already adding an MBR and boot sector to the ISO
image so it can be dd'ed to a USB stick and booted as a hard disk.  For
UEFI systems, we're already generating an ESP and setting it as an
additional El Torito image.  For some UEFI firmware (e.g. EDK II) that
seems to be sufficient, but we should also create a GPT with an ESP
wrapping that El Torito image, for firmware that doesn't support
El Torito on hard disks.

For https://github.com/coreos/fedora-coreos-tracker/issues/707.