Add `cosa buildextend-extensions`
This elevates extensions to be "first-class" artifacts. We include a
tarball of the extensions in the build dir, as well as details in
`meta.json` (this can be used for example by the release browser).

This command will replace the `download-extensions` script. A major
advantage is that it leverages rpm-ostree's new support for extension
composes:

https://github.com/coreos/rpm-ostree/pull/2439

One awkward bit is that our Python code has to call back into the shell
library because we need to make use of the `runvm` path in the
unprivileged case, since that's where the cache is located.