qemuexec: default to no Ignition config
Let's make `qemuexec` slightly less opinionated in the spirit of being
more low-level: if we're not using any fancy features (like bind mounts
or auto-login), let's avoid even generating an empty Ignition config and
pass none.

This helps `kola qemuexec -p qemu-iso` a lot, because (1) it avoids
having to create a secondary ISO only to inject a no-op config (and
wasting developer time doing so), (2) we already have an auto-login
shell by default in the ISO if no Ignition config is provided and
providing an empty one breaks that, and (3) it works around Ignition
config mismatch issues like #1976 (in my case, I'm debugging the 4.6
live ISO).