Add size/md5 to Azure metadata
First this changes ore to write a JSON file that the Python side
parses, so we don't need to re-synthesize the URL on the Python
side.

Gather Azure Blob metadata (size+md5) and add that to our metadata,
This is mainly useful so that one can use the Azure API to "offline validate"
against our metadata snapshot.  I'd also like to add our sha256
checksum for consistency, but that can come later.

This is prep for https://github.com/coreos/stream-metadata-go/issues/13