build-fast: Use .cosa dir, honor in cosa/kola
Having a toplevel file for the `fastbuild-$x.qcow2` clutters
things; let's stick the file in a `.cosa` dir that can be more
easily added to upstream `.gitignore`.

And while we're here, change kola to automatically detect
the presence of a `.qcow2` here if no other option is specified.

This then works:
```
$ cosa build-fast
$ cosa run
```