build.sh: support git.tar.gz in container build dir
The `OPENSHIFT_GIT_HACK` workaround doesn't work well in some situations
because (1) it assumes that there even is a reachable git repo from
which to re-clone the repo; for dev machines using e.g. `--from-dir=.`,
this is not the case, and (2) it introduces a race where the git ref
might have moved forward since the build started, effectively yielding a
src tree with a mix of "old" and "new" files (though it does use `rsync
--ignore-existing` at least).

Specifically, in the `--from-dir=.` workflow, it would be much cleaner
conceptually if it just worked purely from what was copied in and didn't
try to fetch more source code from a remote.

Let's support a convention where if there is a `git.tar.gz` in the
container build directory, we unzip that and delete it.

(I think it wouldn't be too hard either to support building without a
git directory to help this case, though it's definitely useful to have
git information and it's nice to not have to think the missing case.)