kola: Fix localbuild detection
This should fix rpm-ostree CI that was broken by
https://github.com/coreos/coreos-assembler/pull/2040

Basically we need to ignore the empty string (meaning "not found")
as a return value if there's no local build.

Also while we're here, strengthen the check to not even look
for a local build if the qemu disk is already set.