kola: Add --no-default-checks
This will allow us to e.g. bypass SELinux denials when we need
to, e.g.  https://github.com/coreos/fedora-coreos-tracker/issues/751

The intention is that for FCOS releases we would drive this declaratively
in the pipeline, something like a:

```
kola:
  no-default-checks: 33.20210201.3.0
```

to automatically suppress checks for builds with that version number.